% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allignGeometry.R
\docType{methods}
\name{allignGeometry}
\alias{allignGeometry}
\alias{allignGeometry,RasterLayer-method}
\alias{allignGeometry,RasterStack-method}
\alias{allignGeometry,Satellite-method}
\title{Allign raster geometry between two data sets}
\usage{
\S4method{allignGeometry}{Satellite}(x, template, band_codes, type,
  method = "bilinear")

\S4method{allignGeometry}{RasterStack}(x, template, method = "bilinear")

\S4method{allignGeometry}{RasterLayer}(x, template, method = "bilinear")
}
\arguments{
\item{template}{Raster or spatial data set from which geometry can be extracted}

\item{method}{Method for resampling ("ngb" or "bilinear")}

\item{data}{Raster layer to be resampled}
}
\value{
Satellite object with path radiance for each band in the metadata
(W m-2 micrometer-1)

raster::RasterStack object with converted layers

raster::RasterLayer object with converted layer
}
\description{
Allign raster data by bringing it in the same geometry and extend.
If the data set is not in the same projection as the template, the allignment
will be computed by reprojection. If the data has already the same
projection, the data set will be croped and aggregated prior to resampling
in order to reduce computation time.






}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)
allignGeometry(sat, template = getSatDataLayer(sat, "B008n"),
               band_codes = "B001n")
}

