% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcHistMatch.R
\docType{methods}
\name{calcHistMatch}
\alias{calcHistMatch}
\alias{calcHistMatch,RasterLayer-method}
\alias{calcHistMatch,RasterStack-method}
\alias{calcHistMatch,Satellite-method}
\title{Illumination correction across scenes using histogram matching}
\usage{
\S4method{calcHistMatch}{Satellite}(x, target, bcde = NULL, minv = NULL,
  maxv = NULL, step = NULL, ttab = FALSE, use_cpp = TRUE)

\S4method{calcHistMatch}{RasterStack}(x, target, bcde = NULL, minv = NULL,
  maxv = NULL, step = NULL, ttab = FALSE, use_cpp = TRUE)

\S4method{calcHistMatch}{RasterLayer}(x, target, bcde = NULL, minv = NULL,
  maxv = NULL, step = NULL, ttab = FALSE, use_cpp = TRUE)
}
\arguments{
\item{x}{Satellite or raster::Raster* object providing the source band(s) to
be adjusted.}

\item{target}{The target band as raster::RasterLayer.}

\item{minv}{Lower limit of the possible range for transformation (if not
provided, defaults to the minimum of both layers).}

\item{maxv}{Upper limit of the possible range for transformation (if not
provided, defaults to the maximum of both layers).}

\item{ttab}{Logical. If TRUE, the transformation table is being returned.}

\item{use_cpp}{Logical. If \code{TRUE}, C++ functionality (via \strong{Rcpp})
is enabled, which leads to a considerable reduction of both computation time
and memory usage.}

\item{mask}{An invariant feature mask as raster::RasterLayer.}

\item{by}{Step size used to build the new histogram
(if not provided, defaults to 1 for integer master layer and 0.01 for float
master layer).}
}
\value{
If \code{ttab = FALSE} a RasterLayer; \cr
if \code{ttab = TRUE} a list containing
\code{recode} the transformation table used to match the histograms
\code{newraster} the transformed RasterLayer

Satellite object with added atmospheric corrected layers

raster::RasterStack object with atmospheric corrected layers

raster::RasterLayer object with atmospheric corrected layer
}
\description{
This function adjusts the illumination of individual bands across two scenes
using a histogram match.






}
\examples{
## sample data
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

## extraction of source and targed bands
x <- getSatDataLayer(sat, "B004n")
target <- getSatDataLayer(sat, "B005n")

## histogram matching
calcHistMatch(x, target)
}
\references{
This function is taken and only slightly adapted from the
\code{\link{landsat::histmatch}} function by Sarah C. Goslee (2011).
Analyzing Remote Sensing Data in R: The landsat Package. Journal of
Statistical Software, 43(4), 1-25. URL \url{http://www.jstatsoft.org/v43/i04/}.
}

