% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcKappa.R
\name{calcKappa}
\alias{calcKappa}
\title{Calculate various kappa coefficients}
\usage{
calcKappa(ctable)
}
\arguments{
\item{ctable}{contingency table (see function
\code{\link{comContingencyTable}})}
}
\value{
matrix containing different kappa coefficients
}
\description{
Calculate various kappa coefficients for accuracy assessment of
classifications.
}
\details{
currently, the following coefficients are implemented:
standard kappa index
kappa of location
kappa of histogram
chance agreement
quantity agreement
allocation agreement
allocation disagreement
quantity disagreement
}
\examples{
# Not run
calcKappa(contab)
}
\references{
Todo.
}
\seealso{
\code{\link{comContingencyTable}} for compiling a contingency table.
}

