% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcTOAIrradRadRef.R
\name{calcTOAIrradRadRef}
\alias{calcTOAIrradRadRef}
\title{Compute extraterrestrial solar irradiance (ESun) based on Radiation/Reflection}
\usage{
calcTOAIrradRadRef(rad_max, ref_max, esd, normalize = TRUE)
}
\arguments{
\item{rad_max}{maximum radiance of satellite band(s)}

\item{esd}{earth sun distance (AU)}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{rad_min}{minimum radiance of satellite band(s)}
}
\value{
vector object containing ESun for each band
}
\description{
Compute extraterrestrial solar irradiance (ESun) using the actual
maximum radiation and reflection values within each band.
}
\details{
The actual solar irradiance is compute using the following formula
taken from GRASS'
\href{http://grass.osgeo.org/grass65/manuals/i.landsat.toar.html}{i.landsat.toar module}
\deqn{ESun = (pi d^2) RADIANCE_MAXIMUM / REFLECTANCE_MAXIMUM}
where d is the sun-earth distance in astronomical units and RADIANCE_MAXIMUM,
REFLECTANCE_MAXIMUM are the maximum radiance and reflection values of the
respective band. All these parameters are taken from the scene's metadata
file.

By default, the resulting actual ESun will be normalized to a mean earth sun
distance to be compatible with other default results from \code{\link{eSun}}.
The earth sun distance for the normalization is taken from the metadata.
}
\examples{
calcTOAIrradRadRef(coefs = coefs)
}
\seealso{
\code{\link{eSun}} for wrapping this function and alternative
derivation of ESun.
}

