% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calibLinear.R
\docType{methods}
\name{calibLinear}
\alias{calibLinear}
\alias{calibLinear,Raster-method}
\alias{calibLinear,Satellite-method}
\alias{calibLinear,numeric-method}
\title{Convert DNs to radiance, reflectance and/or brightness temperature.}
\usage{
\S4method{calibLinear}{Satellite}(x, convert = "all",
  szen_correction = "TRUE")

\S4method{calibLinear}{Raster}(x, bnbr, mult, add, szen, k1, k2)

\S4method{calibLinear}{numeric}(x, bnbr, mult, add, szen, k1, k2)
}
\arguments{
\item{x}{Satellite, Raster*, or numeric object of the sensor's band values}

\item{bnbr}{number of the band (required if rasterstack)}

\item{coefs}{coefficients data frame resulting from compMetaLandsat()}

\item{conv}{conversion type (one of "rad", "ref", "refsun", "bt")}
}
\value{
Satellite, Raster*, or numeric object object with converted values
}
\description{
The function converts the digital numbers of a Lansat 8 level 1B/T file
to radiance (rad), relfectance (ref) and/or brightness temperature (bt) using
the calibration coefficients from the metadata file. The reflectance
conversion can additionaly be include a sun zenith correction.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

bcde <- "002n"

calibLinear(band = getSatDataLayer(sat, bcde), bnbr = 1,
            mult = getSatRADM(sat, bcde),
            add = getSatRADA(sat, bcde))
}
\references{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\seealso{
\code{\link{radiometricCorrection}} for converions of the DNs
including scene-based atmospheric correction.
}

