% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibLinearInverse.R
\name{calibLinearInverse}
\alias{calibLinearInverse}
\title{Convert reflectance to radiance using linear function coefficients}
\usage{
calibLinearInverse(band, ref_mult, ref_add, rad_mult, rad_add, szen)
}
\arguments{
\item{band}{raster, rasterstack or data frame object of the sensor band}

\item{coefs}{coefficients data frame resulting from compMetaLandsat()}

\item{conv}{conversion type (one of "rad", "ref", "refsun", "bt")}
}
\value{
Raster object with converted values
}
\description{
The function converts the relfectance (ref) back to radiance (rad) given that
linear conversion coefficients for both, radiance and reflectance are
available.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

bcde <- "002n"

calibLinear(band = getSatDataLayer(sat, bcde),
            mult = getSatRADM(sat, bcde),
            add = getSatRADA(sat, bcde))
}
\references{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\seealso{
\code{\link{radiometricCorrection}} for converions of the DNs
including scene-based atmospheric correction.
}

