% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calibrateLandsat.R
\name{calibrateLandsat}
\alias{calibrateLandsat}
\title{Convert DNs to radiance, reflectance and/or brightness temperature.}
\usage{
calibrateLandsat(band, bnbr, coefs, conv = "rad")
}
\arguments{
\item{band}{raster object of the Landsat band}

\item{bnbr}{number of the Landsat band}

\item{coefs}{coefficients data frame resulting from collectLandsat8Metadata()}

\item{conv}{conversion type (one of "rad", "ref", "refsun", "bt")}
}
\value{
Raster object with converted values
}
\description{
The function converts the digital numbers of a Lansat 8 level 1B/T file
to radiance (rad), relfectance (ref) and/or brightness temperature (bt) using
the calibration coefficients from the metadata file. The reflectance
conversion can additionaly be include a sun zenith correction.
}
\examples{
path <- system.file("extdata", package = "satellite")
filesl8 <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
coefs8 <- collectLandsat8Metadata(filesl8)
calibrateLandsat(l8[[2]], 2, coefs8, conv = "ref")
}
\references{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\seealso{
\code{\link{radiometricCorrection}} for converions of the DNs
including scene-based atmospheric correction.
}

