% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compContingencyTable.R
\name{compContingencyTable}
\alias{compContingencyTable}
\title{Compute contingency table}
\usage{
compContingencyTable(pred, ref, ref_id)
}
\arguments{
\item{pred}{raster object containing the predicted values}

\item{ref}{spatial polygon object containing the reference values for certain
locations}

\item{ref_id}{name of the attribute which contains the reference id values}
}
\value{
table containing contingency values
}
\description{
Compute contingency table based on reference and validation data
}
\details{
The predicted values from the raster dataset are extracted if they
fall inside the individual polygons.
}
\examples{
# Not run
calcKappa(contab)
}
\seealso{
\code{\link{calcKappa}} for calculating kappa coefficients based on
the contingency table.
}

