% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compFilePathGLS.R
\name{compFilePathGLS}
\alias{compFilePathGLS}
\title{Get filename, bands and metadata file for Global Land Survey standard 1B/T format}
\usage{
compFilePathGLS(files)
}
\arguments{
\item{files}{Path and filename(s) of one or more GLS band files or, 
alternatively, one or more GLS metadata files.}
}
\value{
\code{data.frame} containing filepaths, band numbers and metadata 
filepaths.
}
\description{
The function compiles the sensor, band, filename and metadata filename information
for standard level 1B/T Global Land Survey (GLS) files.
}
\examples{
\dontrun{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC08*.TIF"), full.names = TRUE)
compFilePathGLS(files)  
}

}
