% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compileFilePathLandsat.R
\name{compileFilePathLandsat}
\alias{compileFilePathLandsat}
\title{Get filename, bands and meta file for Landsat 7 and 8 standard 1B/T format.}
\usage{
compileFilePathLandsat(files)
}
\arguments{
\item{files}{path and filename of one or more landsat band file}
}
\value{
Data frame containing filepathes, band numbers and metafilepathes
}
\description{
The function builds the metadata filename  based on standard level 1B/T
Landsat filenames.
}
\examples{
compileFilePathLandsat(files = c(inst/extdata/LC81950252013188LGN00_B1.tif,
inst/extdata/LC81950252013188LGN00_B10.tif))
}

