% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compMetaLandsat.R
\name{compMetaLandsat}
\alias{compMetaLandsat}
\title{Get calibration information from Landsat 8 standard level 1B/T filename.}
\usage{
compMetaLandsat(files)
}
\arguments{
\item{filepath}{path and filename to the landsat metadata file}
}
\value{
Dataframe containing the following information for each band
\itemize{
  \item Date (DATE), sensor (SENSOR) and band number (BNBR)
  \item addtition and multiplication coefficients for reflectance
  (REFA, REFM) and radiance (RADA, RADM)
  \item brightness temperature correction parameters (BTK1, BTK2)
  \item sun elevation angle (SELEV), sun zenith angle (SZEN) and sun azimuth
  angle (SAZM).
  \item Radiance/reflectance maximum and minimum values (RADMAX, RADMIN,
  REFMAX, REFMIN)
  \item Filepath (FILE) of the sensor data and metadata file (METAFILE)
}
}
\description{
The function scans a Lansat metadata file for various calibration
and orbit coefficients as well as some sensor specific data.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
compMetaLandsat(files)
}

