% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertRef2Rad.R
\name{convertRef2Rad}
\alias{convertRef2Rad}
\title{Convert reflectance to radiance using linear function coefficients}
\usage{
convertRef2Rad(band, refm, refa, radm, rada, szen)
}
\arguments{
\item{band}{raster::RasterStack or raster::RasterLayer containing reflectance}

\item{REFM}{multiplication coefficient for reflectance conversion}

\item{REFA}{addtition coefficient for reflectance conversion}

\item{RADM}{multiplication coefficient for radiance conversion}

\item{RADA}{addtition coefficient for radiance conversion}

\item{SZEN}{sun zenith angle}
}
\value{
Raster object with converted values
}
\description{
The function converts the relfectance (ref) back to radiance (rad) given that
linear conversion coefficients for both, radiance and reflectance are
available.
}
\details{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\examples{
Not run:
}

