% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertSCLinear.R
\docType{methods}
\name{convertSCLinear}
\alias{convertSCLinear}
\alias{convertSCLinear,RasterLayer-method}
\alias{convertSCLinear,RasterStack-method}
\alias{convertSCLinear,Satellite-method}
\title{Convert a band's scaled counts to radiance, reflectance and/or temperature}
\usage{
\S4method{convertSCLinear}{Satellite}(x, convert = "all",
  szen_correction = "TRUE")

\S4method{convertSCLinear}{RasterStack}(x, mult, add, szen, k1, k2)

\S4method{convertSCLinear}{RasterLayer}(x, mult, add, szen, k1, k2)
}
\arguments{
\item{x}{An object of type Satellite,  raster::RasterStack or
raster::RasterLayer}

\item{convert}{Type of physical output, one of "rad", "ref", "bt" or "all"}

\item{mult}{multiplicative coefficient for value transformation (i.e. slope)}

\item{add}{additive coefficient for value transformation (i.e. offset)}

\item{szen}{cosine of solar zenith angle}

\item{k1,k2}{temperature correction parameters}

\item{szen_cor}{Apply sun zenith correction, TRUE or FALSE}

\item{band}{raster, rasterstack or data frame object of the sensor band}
}
\value{
Satellite object with added converted layers

raster::RasterStack object with converted layers

raster::RasterLayer object with converted layer
}
\description{
Convert a band's scalled counts to radiance, reflectance
and or brightness temperature, using a simple linear conversion without
any kind of atmospheric correction etc.






}
\details{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
sat <- convertSCLinear(sat)

# If you use a raster layer, supply required meta information
bcde <- "B002n"
convertSCLinear(band = getSatDataLayer(sat, bcde),
            mult = getSatRADM(sat, bcde),
            add = getSatRADA(sat, bcde))
}
\seealso{
\code{\link{calcAtmosCorr}} for converions of scaled counts
to physical units including a scene-based atmospheric correction.
}

