% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convRad2Ref.R
\docType{methods}
\name{convRad2Ref}
\alias{convRad2Ref}
\alias{convRad2Ref,RasterLayer-method}
\alias{convRad2Ref,RasterStack-method}
\alias{convRad2Ref,Satellite-method}
\title{Convert a band's scaled counts or radiance values to reflectance}
\usage{
\S4method{convRad2Ref}{Satellite}(x, szen_correction = "TRUE")

\S4method{convRad2Ref}{RasterStack}(x, mult, add, szen)

\S4method{convRad2Ref}{RasterLayer}(x, mult, add, szen)
}
\arguments{
\item{x}{An object of type Satellite,  raster::RasterStack or
raster::RasterLayer providing radiance values}

\item{szen_correction}{Apply sun zenith correction, TRUE or FALSE}

\item{mult}{multiplicative coefficient for value transformation (i.e. slope)}

\item{add}{additive coefficient for value transformation (i.e. offset)}

\item{szen}{cosine of solar zenith angle}
}
\value{
Satellite object with added converted layers

raster::RasterStack object with converted layers

raster::RasterLayer object with converted layer
}
\description{
Convert a band's scalled counts to reflectance using a simple linear
conversion without any kind of atmospheric correction etc.






}
\details{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}

If the sensor does not provide linear conversion coefficients for reflectance
computation, the reflectance is calculated using the solar irradiance
following the functions taken from USGS' Landsat 7 manual, chapter 11.3.2,
which is available online at
\url{http://landsathandbook.gsfc.nasa.gov/data_prod/prog_sect11_3.html}
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
sat <- convRad2Ref(sat)

# If you use a raster layer, supply required meta information
bcde <- "B002n"
convRad2Ref(band = getSatDataLayer(sat, bcde),
            mult = getSatRADM(sat, bcde),
            add = getSatRADA(sat, bcde))
}
\seealso{
\code{\link{calcAtmosCorr}} for converions of scaled counts
to physical units including a scene-based atmospheric correction.
}

