% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satCrop.R
\docType{methods}
\name{cropSat}
\alias{cropSat}
\title{Crop satellite object.}
\usage{
\S4method{cropSat}{Satellite}(x, y, subset = TRUE)
}
\arguments{
\item{x}{object of type Satellite}

\item{y}{Extent object}

\item{subset}{logical, defaults to TRUE. Drops all layers but the cropped ones.
If set to false appends cropped layers to Satellite object.}
}
\value{
Satellite object cropped layers
}
\description{
The function is a wrapper around the raster::crop function to easily crop
a satellite object by an extent object.
}
\details{
Crop layers of a Satellite object to the size of a given Extent object.
}
\examples{
#EXAMPLE STILL NEEDS DATA (i.e. the crop template)

path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)
crop_template <- raster(paste0(path,crop_template))

sat <- crop(sat, crop_template)
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for \code{\link[raster]{crop}}.
}

