% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/darkObjectSubstraction.R
\name{darkObjectSubstraction}
\alias{darkObjectSubstraction}
\title{Convert digital numbers to radiances corrected by a dark object substraction}
\usage{
darkObjectSubstraction(sat = 5, scattering.coef = c(-4, -2, -1, -0.7, -0.5),
  SHV, SHV.band, gain, offset, Grescale, Brescale, sunelev, edist,
  Esun = c(198.3, 179.6, 153.6, 103.1, 22, 8.34), blackadjust = 0.01)
}
\arguments{
\item{red}{a raster of the sensor's red band}

\item{nir}{a raster of the sensor's nir band}

\item{swir}{a raster of the sensor's swir band}
}
\value{
raster object with invariant pixels marked with 1, 0 otherwise
}
\description{
Perform a dark object substraction in order to roughly correct for
atmospheric scattering and convert the band's digital numbers to radiance
or reflectance values.
}
\examples{
not run:
darkObjectSubstraction(red, nir, swir, level=.99)
}
\references{
This function is an extended version of the DOS function from
Sarah C. Goslee (2011). Analyzing Remote Sensing Data in R: The landsat
Package. Journal of Statistical Software,43(4), 1-25. URL
\url{http://www.jstatsoft.org/v43/i04/}.

The underlaying theory has been published by Chavez Jr PS (1988) An improved
dark-object subtraction technique for atmospheric scattering correction of
multispectral data. Remote Sensing of Environment 24/3,
doi:10.1016/0034-4257(88)90019-3, available online at
 \url{http://www.sciencedirect.com/science/article/pii/0034425788900193}
}

