% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declination.R
\name{declination}
\alias{declination}
\title{Compute declination angle}
\usage{
declination(n, formula = c("Cooper", "Spencer"))
}
\arguments{
\item{n}{Day of the year (DOY) as \code{integer} or \code{Date} from which to
derive DOY.}

\item{formula}{Formula to be applied, specified through the name of the 
author (see References). Currently, the approximation equation of Cooper 
(1969) and the more accurate estimation of Spencer (1971) are implemented.}
}
\value{
A \code{numeric} declination angle.
}
\description{
Compute the declination of the sun at solar noon based on the day of the 
year.
}
\details{
"The declination of the sun is the angle between the equator and a line drawn 
from the centre of the Earth to the centre of the sun." (directly taken from 
\url{http://www.pveducation.org/pvcdrom/properties-of-sunlight/declination-angle})
}
\examples{
declination(Sys.Date())
declination(Sys.Date(), formula = "Spencer")

}
\references{
The formulas are taken from the following sources:
\itemize{
  \item \strong{Cooper}: Cooper PI (1969) The absorption of radiation in 
  solar stills. Solar Energy 12(3), 333--346, 
  \url{http://www.sciencedirect.com/science/article/pii/0038092X69900474}.
  \item \strong{Spencer}: Spencer JW (1971) Fourier series representation of 
  the position of the sun. Search 2(5), 172, \url{https://goo.gl/lhi9UI}.
}
See also 
\itemize{
  \item Duffie JA, Beckman WA (2013) Solar Engineering of Thermal Processes. 
  Wiley: Hoboken, New Jersey, ISBN: 978-0-470-87366-3,
  \url{http://eu.wiley.com/WileyCDA/WileyTitle/productCd-0470873663.html} and 
  \item ITACA (2016) Part 3: Calculating Solar Angles, 
  \url{http://www.itacanet.org/the-sun-as-a-source-of-energy/part-3-calculating-solar-angles/}
}
for further information.
}

