% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/depricated.R
\name{depricated}
\alias{calibLinear}
\alias{depricated}
\alias{satAtmosCorr}
\alias{satCalib}
\alias{satPathRadDOS}
\alias{satTOAIrrad}
\title{Depricated functions}
\usage{
satCalib(x, convert = "all", szen_correction = "TRUE")

calibLinear(band, mult, add, szen, k1, k2)

satTOAIrrad(x, method = "Table", model = "MNewKur", normalize = TRUE, esd)

satPathRadDOS(x, atmos_model = "DOS2", esun_mode = "RadRef")

satAtmosCorr(x, atmos_model = "DOS2", esun_mode = "RadRef")
}
\description{
The functions have been implemented in the very beginging of the package
development, mainly to be used within a course on remote sensing at
Marburg University. To ensure that the scripts developed within this course
will still work after the next major revision, they are still part of this
package but they will mainly just foreward the respective call to the
more up-to-date function.
}

