% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\docType{methods}
\name{extend}
\alias{extend}
\alias{extend,Satellite-method}
\title{extend satellite object}
\usage{
\S4method{extend}{Satellite}(x, y, subset = TRUE)
}
\arguments{
\item{x}{Satellite object.}

\item{y}{\code{\link{extent}} object.}

\item{subset}{Logical; if \code{TRUE} (default), all layers but the extendped 
ones are being dropped; if \code{FALSE}, extendped layers are appended to the 
Satellite object.}
}
\value{
A Satellite object consisting of extended layers only. If 
\code{subset = FALSE}, a Satellite object with the extendped layers appended.
}
\description{
The function is a wrapper around the \code{\link{extend}} function to 
easily extend a Satellite object by an \code{\link{extent}} object.
}
\details{
extends layers of a Satellite object to the size of a given 
\code{raster::extent} object.
}
\examples{
\dontrun{
## sample data
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC08*.TIF"), full.names = TRUE)
sat <- satellite(files)

## geographic extent of georg-gassmann-stadium (utm 32-n)
ext_ggs <- raster::extent(484015, 484143, 5627835, 5628020)

## extend satellite object by specified extent
sat_ggs <- extend(sat, ext_ggs)

plot(sat)
plot(sat_ggs)
}
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for \code{raster::extend}.
}
