% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSolarZenith.R
\name{fromSolarZenith}
\alias{fromSolarZenith}
\title{Compute hour angle and time from the solar zenith angle}
\usage{
fromSolarZenith(theta, delta, phi)
}
\arguments{
\item{theta}{\code{numeric} sun zenith angle in degrees (see Note).}

\item{delta}{\code{numeric} angular declination of the sun in degrees.}

\item{phi}{\code{numeric} latitude in EPSG:4326 (see 
\url{http://spatialreference.org/ref/epsg/wgs-84/}).}
}
\description{
Compute the solar hour angle and the time difference from noon associated 
therewith from the sun zenith angle.
}
\note{
Negative and positive solar hour angles indicate solar zenith angles during 
the morning and afternoon, respectively. Note that the sun zenith angle can 
easily be derived from the elevation angle through 
\code{sun zenith = 1 - elevation}.
}
\examples{
theta <- 30 # 30 deg zenith angle
delta <- declination(Sys.Date(), "Spencer", "degrees") # angular declination
phi <- 0    # geographic latitude, i.e. equator

fromSolarZenith(theta, delta, phi) 

}
\references{
PVEducation (2016) Elevation angle. Available online at 
\url{http://www.pveducation.org/pvcdrom/properties-of-sunlight/elevation-angle}.
}
\seealso{
\code{\link[satellite]{declination}}, \code{\link{solarHourAngle}}.
}

