% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLandsatSceneCoef.R
\name{getSceneCoef}
\alias{getSceneCoef}
\title{Get calibration information from Landsat 8 standard level 1B/T filename.}
\usage{
getSceneCoef(filepath, band, coef = "rad")
}
\arguments{
\item{filepath}{path and filename to the landsat metadata file}

\item{band}{band number for which the coefficients should be retrieved}

\item{coef}{either "rad" or "ref" for radiance/reflectance coefficients}
}
\value{
vector containing, multiplication coefficient [1],
addition coefficient [2], sun elevation angle [3], sun zenith angle [4],
sun azimuth angle [5]
}
\description{
The function scans a Lansat 8 metadata file for calibration coefficients and
solar geometry information. Depending on the user selection, the calibration
coefficients are for reflectance or radiance.
}
\examples{
not run:
getSceneCoef(filepath = "Name_of_Landsat_Metadata_File", band = 5)
}

