% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exwin.R
\name{incl}
\alias{incl}
\title{Compute similarity index}
\usage{
incl(x)
}
\arguments{
\item{pred}{raster object containing the predicted values}

\item{ref}{spatial polygon object containing the reference values for certain
locations}

\item{win}{max window size. For now defaults to 3}

\item{weights}{vector with weights for layers of each window}
}
\value{
similarity index
}
\description{
Compute similarity index based on expanding window approach
}
\details{
Similarity index based on the approach in \code{\link{http://www.likbez.com/AV/PUBS/Kuhnert2005PERS.pdf}}.
}
\examples{
not run:
exwin(pred_raster, training_sites_shape)
}
\seealso{
\code{\link{compContingencyTable}} \code{\link{calcKappa}} for calculating kappa coefficients based on
contingency tables.
}

