% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maskInvarFeatures.R
\name{invariantFeatures}
\alias{invariantFeatures}
\title{Identify pseudi-invariant features from a satellite scene}
\usage{
invariantFeatures(red, nir, swir, level = 0.99)
}
\arguments{
\item{red}{a raster of the sensor's red band}

\item{nir}{a raster of the sensor's nir band}

\item{swir}{a raster of the sensor's swir band}
}
\value{
raster object with invariant pixels marked with 1, 0 otherwise
}
\description{
Identify pseudi-invariant features from a satellite scene based on a
red, near-infrared and short-wave infrared band.
}
\examples{
not run:
invariantFeatures(red, nir, swir, level=.99)
}
\references{
This function is taken and only slightly adapted from the PIF
function of Sarah C. Goslee (2011). Analyzing Remote Sensing Data in R: The
landsat Package. Journal of Statistical Software,43(4), 1-25. URL
\url{http://www.jstatsoft.org/v43/i04/}.

The underlaying theory has been published by Schott RJ, Salvaggio C and
Volchok WJ (1988) Radiometric scene normalization using pseudoinvariant
features. Remote Sensing of Environment 26/1,
doi:10.1016/0034-4257(88)90116-2, available online at
 \url{http://www.sciencedirect.com/science/article/pii/0034425788901162}
}

