% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite-package.R
\docType{data}
\name{l5}
\alias{l5}
\title{Landsat 5 sample data}
\format{\code{raster::RasterStack} with 8 bands (Landsat 5 bands 1 to 7 plus 
quality layer) of 101 by 101 pixels.}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\description{
This dataset comes from the USGS. It contains part of the 
Landsat 5 scene LT05_L1TP_167055_20000309_20161214_01_T1 from 2000-03-09 over 
the Bale Mountains, Ethiopia.
}
\details{
Use of this data requires your agreement to the USGS regulations on 
using Landsat data.
}

