% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_l7.R
\docType{data}
\name{l7}
\alias{l7}
\title{Landast 7 raster stack.}
\format{RasterLayer with 8 bands of 41 by 41 pixels.}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\usage{
l7
}
\description{
This dataset comes from the USGS. It contains part of the Landsat scene
LE71950252001211EDC00 from 2001-07-30 over Maburg, Germany.
}
\section{Variables}{

\describe{
\item{band}{Landsat 7 bands}
}
}

\section{Use}{

Use of this data requires your agreement to the USGS regulations on using
Landat data.
}
\keyword{datasets}

