% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_l8.R
\docType{data}
\name{l8}
\alias{l8}
\title{Landast 8 raster stack.}
\format{RasterLayer with 10 bands of 41 by 41 pixels.}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\usage{
l8
}
\description{
This dataset comes from the USGS. It contains part of the Landsat scene
LC81950252013188LGN00 from 2013-07-07 over Maburg, Germany.
}
\section{Variables}{

\describe{
\item{band}{Landsat 8 bands}
}
}

\section{Use}{

Use of this data requires your agreement to the USGS regulations on using
Landat data.
}
\keyword{datasets}

