% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metaFilePathLandsat.R
\name{metaFilePathLandsat}
\alias{metaFilePathLandsat}
\title{Get metadata filename from Landsat 8 standard level 1B/T filename.}
\usage{
metaFilePathLandsat(filepath)
}
\arguments{
\item{filepath}{path and filename to the landsat band file (not the metadata)}
}
\value{
List containing metadata filepath [[1]] and band number [[2]]
}
\description{
The function builds the metadata filename  based on standard level 1B/T
Landsat filenames.
}
\examples{
landsat8_filepath <- "LC81950252013188LGN00_B2.TIF"
metaFilePathLandsat(filepath = landsat8_filepath)
}

