% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mocRS-package.R
\docType{package}
\name{mocRS-package}
\alias{mocRM}
\alias{mocRS-package}
\title{Smorgasboard for remote sensing functions}
\description{
Smorgasboard for remote sensing functions
}
\details{
The package provides a variety of functions which are usefull for learning
log assignments as part of the MSC Environmental Geography course on
remote sensing at www.moc.environmentalinformatics-marburg.de.
}
\author{
Thomas Nauss\cr
\cr
\emph{Maintainer:} Environmental Informatics \email{admin@environmentalinformatics-marburg.de}
}
\keyword{package}

