% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,Satellite-method}
\alias{names,Satellite-method}
\alias{names<-,Satellite-method}
\title{Get/set Satellite data layer names}
\usage{
\S4method{names}{Satellite}(x)

\S4method{names}{Satellite}(x) <- value
}
\arguments{
\item{x}{A Satellite object.}

\item{value}{Band codes of the individual data layers.}
}
\value{
Satellite data layer names as character vector.
}
\description{
Get/set Satellite data layer names, i.e. the BCDE id.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC08*.TIF"), full.names = TRUE)
sat <- satellite(files)
names(sat)
new_names <- paste0(names(sat), "_test")
names(sat) <- new_names
}
