% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathRadianceDOS.R
\name{pathRadianceDOS}
\alias{pathRadianceDOS}
\title{Compute path radiance based on dark object method}
\usage{
pathRadianceDOS(sensor = "Landsat 8", DNmin = 69, bnbr = 1, coefs, date,
  scat_coefs = c(-4, -2, -1, -0.7, -0.5), dos_adjust = 0.01)
}
\arguments{
\item{red}{a raster of the sensor's red band}

\item{nir}{a raster of the sensor's nir band}

\item{swir}{a raster of the sensor's swir band}
}
\value{
raster object with invariant pixels marked with 1, 0 otherwise
}
\description{
Compute an estimaed path radiance for all sensor bands using a dark object
method which can be used to roughly correct the radiance values for
atmospheric scattering.
}
\details{
The dark object substraction approach is based on an approximation
of the atmospheric path radiance (i.e. upwelling radiation which is
scatter into the sensors field of view, aka haze) using the reflectance of a
dark object (i.e. reflectance ~1%).

Chavez Jr PS (1988) proposed a method which uses the dark object reflectance
in one band ot predict the corresponding path radiances in all other bands
using a relative radiance model which depends on the wavlelength and overall
atmospheric optical thickness (which is estimated based on the dark object's
DN value). This has the advantage that the path radiance is actually
correlated across different sensor bands and not computed individuall for
each band using independent dark objects.

The relative radiance model follows a wavelength dependent scattering which
ranges from -4 power over -2, -1, -0.7 to -0.5 power for very clear over
clear, moderate, hazy to very hazy conditions. The relative factors are
computed individually for each 1/1000 wavelength within each band range
and subsequently averaged over the band.
}
\examples{
landsat8_metadatafile <-   system.file("extdata",
  "LC81950252013188LGN00_MTL.txt", package = "satellite")
  coefs8 <- landsatCoefficients(landsat8_metadatafile)
  pathRadianceDOS(sensor = "Landsat 8", DNmin = min(raster::getValues(l8)),
  bnbr = 1, coefs = coefs8, date = "2013-07-30")
}
\references{
This function is an extended version of the DOS function from
Sarah C. Goslee (2011) Analyzing Remote Sensing Data in R: The landsat
Package. Journal of Statistical Software,43/4, 1-25. URL
\url{http://www.jstatsoft.org/v43/i04/}.

The underlaying theory has been published by Chavez Jr PS (1988) An improved
dark-object subtraction technique for atmospheric scattering correction of
multispectral data. Remote Sensing of Environment 24/3,
doi:10.1016/0034-4257(88)90019-3, available online at
 \url{http://www.sciencedirect.com/science/article/pii/0034425788900193}.

If you refer to Sawyer and Stephen 2014, please note that eq. 5 is wrong.
}

