% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,Satellite,ANY-method}
\title{Plot a 'Satellite' object}
\usage{
\S4method{plot}{Satellite,ANY}(x, bcde = NULL, ...)
}
\arguments{
\item{x}{A 'Satellite' object, usually returned by \code{\link{satellite}}.}

\item{bcde}{Band codes to be visualized. If not supplied, the initial (up to)
16 layers are being visualized.}

\item{...}{Further arguments passed on to \code{\link{plot}}.}
}
\description{
This is the standard plotting routine for the 'Satellite' class. Layers are
drawn either from the start (default; limited to a maximum of 16 sub-plots
similar to \code{\link{raster::plot}}) or according to the speficied band
codes.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

sat_ps <- panSharp(sat)
}

