% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psEhlers.R
\docType{methods}
\name{psEhlers}
\alias{psEhlers}
\alias{psEhlers,RasterStack-method}
\alias{psEhlers,Satellite-method}
\title{Pan sharpen low resolution satellite channels by using the high resolution 
panchromatic channel.}
\usage{
\S4method{psEhlers}{Satellite}(x, res.method = "ngb", filter = list(win =
  "Han", frq.lowpass, fr.highpass), padzero = FALSE, subset = FALSE)

\S4method{psEhlers}{RasterStack}(x, PAN, res.method = "ngb",
  filter = list(win = "Han", frq.lowpass, fr.highpass), padzero = FALSE,
  subset = FALSE)
}
\arguments{
\item{x}{Satellite or \code{raster::Raster*} object.}

\item{res.method}{resampling method to be used for the xs channels. Currently methods of
\code{\link[raster]{resample}}, namely \code{"ngb"} and \code{"bilinear"} are allowed.}

\item{filter}{list object for defining filter window type and further filter parameters.
Currently only Han window is implemented and cut of frequency by default is calculated
by resolution ratio of xs to PAN resolution.}

\item{padzero}{Logical; defaults to \code{FALSE}. Option to zero pad images before applying FFT.
Currently only images with even number of rows and columns can be zero padded. If zero padding
is choosen images need to be cropped to have even numer of rows and columns (see
\code{\link[satellite]{crop}}). By default if number of rows/ columns are uneven but padding
ist set to \code{TRUE} function will warn but continue without padding.
For information on zero padding see for example Butz (2011) in References.}

\item{subset}{Logical; if TRUE, all layers except for the cropped ones are being dropped;
if FALSE, the cropped layers are being appended to the Satellite object.}

\item{PAN}{A raster::RasterLayer object of the panchromatic channel.}
}
\value{
If x is a Satellite object, a Satellite object (with added 
pansharpened layers); if x is a \code{raster::stack*} object, a 
\code{raster::stack*} with pansharpened layer(s).
}
\description{
Pan sharpen low resolution satellite channels by using the high resolution 
panchromatic channel using a combination of IHS transformation and 
fourier filtering as proposed by Ehlers (2004).
}
\details{
This function implements the PAN sharpening algorithm described in Klonus & Ehlers (2007)
and Ling et. al. (2007). It sharpens satellite images by using a combination of IHS transformation
and subsequent filtering of the Intensity component in the frequency domain. After filtering of the
intensity components of PAN and XS the images are reverse FFT transformed and summed up to form the 
new intensity component to be used for the reverse IHS transformation, which results in the new
pansharpened RGB channels. For more detailed description see references.
}
\examples{


\dontrun{
}
}
\references{
Ehlers, M., 2004: Spectral characteristics preserving image fusion based on Fourier domain filtering. Vol. 5574 of, 1–13 http://dx.doi.org/10.1117/12.565160 (Accessed July 26, 2015).

Klonus, S., and M. Ehlers, 2007: Image Fusion Using the Ehlers Spectral Characteristics Preservation Algorithm. GIScience & Remote Sensing, 44, 93–116, doi:10.2747/1548-1603.44.2.93.

Ling, Y., M. Ehlers, E. L. Usery, and M. Madden, 2007: FFT-enhanced IHS transform method for fusing high-resolution satellite images. ISPRS Journal of Photogrammetry and Remote Sensing, 61, 381–392, doi:10.1016/j.isprsjprs.2006.11.002.

Fisher, R, et. al. 2000: HYPERMEDIA IMAGE PROCESSING REFERENCE. \url{http://homepages.inf.ed.ac.uk/rbf/HIPR2/fourier.htm}

Weinhaus, F., 2011: ImageMagick v6 Examples -- Fourier Transforms. \url{http://www.imagemagick.org/Usage/fourier/#im_fft}

Butz, T., 2011: Fouriertransformation für Fußgänger. 7., aktualisierte Aufl. Vieweg + Teubner, Wiesbaden. (German)
\url{}
}

