% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiometricCorrection.R
\name{radiometricCorrection}
\alias{radiometricCorrection}
\title{Radiometric correction of remote sensing data}
\usage{
radiometricCorrection(x, gain, offset, Grescale, Brescale, sunelev,
  satzenith = 0, edist, Esun, Lhaze, method = "apparentreflectance")
}
\arguments{
\item{filepath}{path and filename to the landsat metadata file}
}
\value{
Raster object containing converted data.
}
\description{
The function computes an atmospheric scattering correction and converts
the sensors digital numbers to radiances. One of four models can be choosen:
- absolute radiance correction \cr
- dark object substraction (Chaves 1989)
- costz (Chaves 1996)
- dark object substratcion (Song et al. 2001)
}
\examples{
not run:
radiometricCorrection(filepath = "Name_of_Landsat_Metadata_File")
}
\references{
This function is an adapted version of the radiocorr function
from Sarah C. Goslee (2011). Analyzing Remote Sensing Data in R: The landsat
Package. Journal of Statistical Software,43(4), 1-25. URL
\url{http://www.jstatsoft.org/v43/i04/}.

The underlaying theory has been published by the following authors:

Chavez Jr PS (1988) An improved dark-object subtraction technique for
atmospheric scattering correction of multispectral data. Remote Sensing of
Environment 24/3, doi:10.1016/0034-4257(88)90019-3, available online at
 \url{http://www.sciencedirect.com/science/article/pii/0034425788900193}

Chavez Jr PS (1996) Image-based atmospheric corrections revisited and
improved. Photogrammetric Engineering and Remote Sensing 62/9,
available online at
\url{http://www.asprs.org/PE-RS-Journals-1996/PE-RS-September-1996.html}

Song C, Woodcock CE, Seto KC, Lenney MP, Macomber SA (2001) Classification
and Change Detection Using Landsat TM Data: When and How to Correct
Atmospheric Effects? Remote Sensing of Environment 75/2,
doi:10.1016/S0034-4257(00)00169-3, available online at
\url{http://www.sciencedirect.com/science/article/pii/S0034425700001693}
}

