% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satAtmosCorr.R
\docType{methods}
\name{satAtmosCorr}
\alias{satAtmosCorr}
\title{Calculate an atmospheric correction for the layers of a Satellite object}
\usage{
\S4method{satAtmosCorr}{Satellite}(x, atmos_model = "DOS2",
  esun_mode = "RadRef")
}
\arguments{
\item{x}{object of type Satellite}

\item{model}{model model to be used (DOS2, DOS4)}

\item{calc_esun}{method used for computation of eSun if not already part of
the Satellite object's metadata (Table, Model, RadRef)}

\item{meta_param}{data frame containing the name of the metadata ID field
used to merge with the readily existing metadata information of the
Satellite class object in the first column and with the same column header
as used in the existing metadata information. All other columns contain the
information to be added to the metadata.}
}
\value{
Satellite object with added, atmospheric corrected layers
}
\description{
The function computes an atmospheric scattering correction and converts
the sensors digital numbers to reflectances using
- absolute radiance correction \cr
- DOS2: a dark object substraction model by Chavez (1996)
- DOS4: a dark object substratcion model by Moran et al. (1992)

If not passed to the function, the necessary estimates of the TOA solar
irradiance are computed, too.
}
\details{
Atmospheric correction is computed using \code{\link{calcAtmosCorr}}.

If necessary (i.e. Satellite object does not contain the respective
information) the following is computed before the actual atmospheric
correction:

The TOA solar irradiance is computed using \code{\link{satTOAIrrad}}.

The path radiance is computed based on a dark object's DN using
\code{\link{calcPathRadDOS}}.

The bands' DNs are converted to radiance using \code{\link{satCalib}}.

Please refer to the respective functions for details on the computation.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)
satTOAIrradTable(sat)
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for
\code{\link{satTOAIrrad}} which is used to get TOA solar irradiance if
nesessary and \code{\link{calcAtmosCorr}} which corrects the sensor data for
atmospheric influences.
}

