% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satCalib.R
\docType{methods}
\name{satCalib}
\alias{satCalib}
\title{Calibrate SC of Satellite object to radiance, reflectance and/or temperature}
\usage{
\S4method{satCalib}{Satellite}(x, convert = "all", szen_correction = "TRUE")
}
\arguments{
\item{x}{object of type Satellite}

\item{method}{name of the method to be used ("Table", "Model", "RadRef)}

\item{model}{name of the model to be used if method is "Model"
(see \code{\link{calcTOAIrradModel}})}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{esd}{earth-sun distance (AU), if not supplied and necessary for
normalize, it is tried to take it from the metadata, otherwise it is estimated
by the day of the year using \code{\link{calcEartSunDist}}.}
}
\value{
Satellite object with added data layers for the respective conversions
}
\description{
Calibrate the scalled counts of a Satellite object to radiance, reflectance
and or brightness temperature, using standard calibration procedures without
atmospheric correction etc.
}
\details{
The calibration is computed using \code{\link{calibLinear}}.

Please refer to the respective functions for details on the computation.
}
\examples{
satPathRadDOSModel(sat)
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
satPathRadDOSRadRef(sat)
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for \code{\link{calibLinear}}.
}

