% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satConvertLinear.R
\docType{methods}
\name{satConvertLinear}
\alias{satConvertLinear}
\title{Convert a band's scaled counts to radiance, reflectance and/or temperature}
\usage{
\S4method{satConvertLinear}{Satellite}(x, convert = "all",
  szen_correction = "TRUE")
}
\arguments{
\item{x}{An object of type Satellite or a raster::RasterStack}

\item{convert}{Type of physical output, one of "rad", "ref", "bt" or "all"}

\item{szen_cor}{Apply sun zenith correction, TRUE or FALSE}
}
\value{
Satellite object with added converted layers
}
\description{
Convert a band's scalled counts to radiance, reflectance
and or brightness temperature, using a simple linear conversion without
any kind of atmospheric correction etc.
}
\details{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
sat <- satConvertLinear(sat)
}
\seealso{
\code{\link{calcAtmosCorr}} for converions of scaled counts
to physical units including a scene-based atmospheric correction.
}

