% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAAsatelliteClasses.R
\docType{class}
\name{Satellite-class}
\alias{Satellite-class}
\title{An S4 class to represent a complete satellite dataset}
\description{
An S4 class to represent a complete satellite dataset
}

