% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satellite-package.R
\docType{package}
\name{satellite-package}
\alias{satellite}
\alias{satellite-package}
\title{Smorgasboard for remote sensing functions}
\description{
Smorgasboard for remote sensing functions
}
\details{
The package provides a variety of functions which are usefull for
handling, manipulating and visualizing remote sensing data.
}
\author{
Thomas Nauss, Hanna Meyer\cr
\cr
\emph{Maintainer:} Environmental Informatics \email{admin@environmentalinformatics-marburg.de}
}
\references{
Some functions are taken and/or addopted from Sarah C. Goslee (2011). Analyzing Remote Sensing Data in R: The landsat Package. Journal of Statistical Software,43(4), 1-25. URL \url{http://www.jstatsoft.org/v43/i04/}.
}
\keyword{package}

