% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.R
\name{satellite}
\alias{satellite}
\alias{satellite,character-method}
\alias{satellite,Raster-method}
\alias{satellite,list-method}
\title{Create a Satellite object}
\usage{
\S4method{satellite}{character}(x, meta, log)

\S4method{satellite}{Raster}(x, meta, log)

\S4method{satellite}{list}(x, meta, log)
}
\arguments{
\item{x}{A vector of filenames, a (multi-layered) \code{Raster*} object or a 
\code{list} of single \code{RasterLayer} objects (see \code{\link[raster]{raster}}). 
In the latter case, be aware that bands must be arranged in ascending order 
(eg using \code{\link{sortFilesLandsat}}).}

\item{meta}{Optional metadata object (e.g. returned from 
\code{\link{compMetaLandsat}}). If 'x' is a satellite dataset and recognised
as "Landsat", then the metadata is automatically extracted from the 
respective meta information file if both the satellite data and the metadata 
file follow the USGS Earth Explorer's naming convention.}

\item{log}{Optionally supply a log entry.}
}
\value{
A \code{Satellite} object.
}
\description{
Method to create a Satellite object.
}
\details{
A Satellite object consists of three data sections:
(i) a raster data section which holds the actual data values of the 
respective sensor bands, (ii) a metadata grid which holds meta information
for each sensor band (e.g. calibration coefficients, type of sensor band 
etc.) and (iii) a list of log information which records the processing 
history of the entire dataset.
}
\examples{
## 'character' input (i.e. filenames)
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC08*.TIF"), full.names = TRUE)

satellite(files)

## raster::RasterStack input
satellite(l8)

}
\seealso{
(i) \code{\link{compMetaLandsat}} to get more information about the
structure of the metadata component; (ii)
\url{https://www.usgs.gov/faqs/what-naming-convention-landsat-collections-level-1-scenes?qt-news_science_products=0#qt-news_science_products} 
for detailed information about the naming conventions for Landsat scene 
identifiers; and (iii) \code{\link{sortFilesLandsat}} for automated 
rearrangement of Landsat band files.
}
