% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AAAsatelliteClasses.R
\docType{class}
\name{SatelliteData-class}
\alias{SatelliteData-class}
\title{An S4 class to represent satellite data}
\description{
An S4 class to represent satellite data
}
\section{Slots}{

\describe{
\item{\code{data}}{a data.frame object containing the metadata}
}}

