% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAsatelliteClasses.R
\docType{class}
\name{SatelliteInfo-class}
\alias{SatelliteInfo-class}
\title{An S4 class to represent a satellite data file}
\description{
An S4 class to represent a satellite data file
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the data file without extension}

\item{\code{filepath}}{full path and file of the data file}

\item{\code{path}}{path to the data file}

\item{\code{file}}{filename incl. extension of the data file}

\item{\code{extension}}{extension of the data file}
}}

