% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAsatelliteClasses.R
\docType{class}
\name{SatelliteLayers-class}
\alias{SatelliteLayers-class}
\title{An S4 class to represent satellite data}
\description{
An S4 class to represent satellite data
}
\section{Slots}{

\describe{
\item{\code{layers}}{a list object containing individual RasterLayer objects}
}}

