% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAsatelliteClasses.R
\docType{class}
\name{SatelliteLog-class}
\alias{SatelliteLog-class}
\title{An S4 class to represent satellite log data}
\description{
An S4 class to represent satellite log data
}
\section{Slots}{

\describe{
\item{\code{log}}{a list object containing information on individual processing steps}
}}

