% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAAsatelliteClasses.R
\docType{class}
\name{SatelliteMetaData-class}
\alias{SatelliteMetaData-class}
\title{An S4 class to represent satellite metadata}
\description{
An S4 class to represent satellite metadata
}
\section{Slots}{

\describe{
\item{\code{meta}}{a data frame object containing the data}
}}

