% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satInvarFeatures.R
\docType{methods}
\name{satInvarFeatures}
\alias{satInvarFeatures}
\title{Identify pseudi-invariant features for a Satellite object}
\usage{
\S4method{satInvarFeatures}{Satellite}(x)
}
\arguments{
\item{x}{object of type Satellite}
}
\value{
Satellite object with added mask
}
\description{
Identify pseudi-invariant features from a satellite scene based on a
vis, near-infravis and short-wave infravis band.
}
\details{
The invariant features are computed using \code{\link{maskInvarFeatures}}.

Please refer to the respective function for details on the computation.
}
\examples{
satPathRadDOSModel(sat)
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
satInvarFeatures(sat)
}
\references{
Please refer to the respective function for references.
}
\seealso{
This function is a wrapper for \code{\link{maskInvarFeatures}}.
}

