% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satPanSharpening.R
\docType{methods}
\name{satPanSharp}
\alias{satPanSharp}
\title{Pan sharpen low resolution satellite channels by using the high resolution panchromatic channel.}
\usage{
\S4method{satPanSharp}{Satellite}(x, filter = "mean", winsize = 3,
  subset = FALSE)
}
\arguments{
\item{x}{Object of type Satellite}

\item{filter}{type of filter to be used for smoothing the PAN raster (e.g. mean (default), Gauss, median).}

\item{winsize}{with n integer size of filter window n x n in pixels. Defaults to 3.
#}

\item{subset}{logical, defaults to TRUE. Drops all layers but the cropped ones.
If set to false appends cropped layers to Satellite object.}
}
\value{
Satellite object
}
\description{
The function PAN sharpens the low resolution channels with the pachromatic channel.
This is done by multiplying the normlized XS channel with the PAN channel (see Details).
}
\details{
This is a wrapper function for panSharp applicable to satellite objects. See \code{\link{panSharp}}
for details.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
satPanSharp(sat)
}

