% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satPathRadDOS.R
\docType{methods}
\name{satPathRadDOS}
\alias{satPathRadDOS}
\title{Compute path radiance based on dark object method for a Satellite object}
\usage{
\S4method{satPathRadDOS}{Satellite}(x, atmos_model = "DOS2",
  esun_mode = "RadRef")
}
\arguments{
\item{x}{object of type Satellite}

\item{method}{name of the method to be used ("Table", "Model", "RadRef)}

\item{model}{name of the model to be used if method is "Model"
(see \code{\link{calcTOAIrradModel}})}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{esd}{earth-sun distance (AU), if not supplied and necessary for
normalize, it is tried to take it from the metadata, otherwise it is estimated
by the day of the year using \code{\link{calcEartSunDist}}.}
}
\value{
Satellite object with path radiance for each band in the metadata
}
\description{
Compute an estimaed path radiance for all sensor band_wls using a dark object
method which can be used to roughly correct the radiance values for
atmospheric scattering.
}
\details{
The path radiance is computed using \code{\link{calcPathRadDOS}}.

If the TOA solar irradiance is not part of the metadata of the Satellite
object, it is computed using \code{\link{satTOAIrrad}}.

Please refer to the respective functions for details on the computation.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)
satPathRadDOSTable(sat)

path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

satPathRadDOSModel(sat)
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
satPathRadDOSRadRef(sat)
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for \code{\link{calcPathRadDOS}}.
}

