% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satSubset.R
\docType{methods}
\name{satSubset}
\alias{satSubset}
\title{Subset a Satellite object}
\usage{
\S4method{satSubset}{Satellite}(x, subset)
}
\arguments{
\item{x}{A Satellite object.}

\item{subset}{Name of the computed instance (e.g. SC, cropped, REF_AtmosCorr
etc.) or row numbers of layers to be selected.}
}
\value{
A Satellite object.
}
\description{
Method to subset a Satellite object. \cr
IMPORTANT: if the name of the target variable where subset is to be stored is
the same as the old one all other information will be lost!
}
\details{
Subsetting of computation stages.
}
\examples{
## sample data
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)

## geographic extent of georg-gassmann-stadium (utm 32-n)
ext_ggs <- raster::extent(484015, 484143, 5627835, 5628020)

## crop satellite object by specified extent
sat <- cropSat(sat, ext_ggs, subset = FALSE)
length(attr(sat, "layers"))

## select cropped layers only
sat_sub <- satSubset(sat, subset = "cropped")
length(attr(sat_sub, "layers"))
}

