% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarHourAngle.R
\docType{methods}
\name{solarHourAngle}
\alias{solarHourAngle}
\alias{solarHourAngle,POSIXt-method}
\alias{solarHourAngle,Raster-method}
\alias{solarHourAngle,character-method}
\title{Compute solar hour angle}
\usage{
\S4method{solarHourAngle}{POSIXt}(x, unit = c("radians", "degrees"))

\S4method{solarHourAngle}{character}(x, unit = c("radians", "degrees"), ...)

\S4method{solarHourAngle}{Raster}(x, origin = strptime("1993-01-01",
  "\%Y-\%m-\%d"), unit = c("radians", "degrees"), ...)
}
\arguments{
\item{x}{A date-time object of class \code{POSIX*}, or a \code{character} 
object that can be coerced by \code{\link{strptime}}, or a \code{Raster*} 
object holding such information.}

\item{unit}{\code{character}, determines whether to return the hour angle in 
radians (default) or degrees.}

\item{...}{If \code{x} is a \code{character} object, formatting arguments 
passed on to \code{\link{strptime}}, or if \code{x} is a \code{Raster*} 
object, formatting arguments related with the specified \code{origin}.}

\item{origin}{A \code{POSIX*} or \code{character} object that defines the 
starting point of \code{x}, defaults to the origin of MOD05_L2 
'Scan_Start_Time' (midnight of 1 January 1993, see 
\url{http://modis-atmos.gsfc.nasa.gov/MOD05_L2/format.html}).}
}
\value{
\code{numeric} solar hour angle(s), or if \code{x} is a \code{Raster*} 
object, a \code{Raster*} object with pixel-wise solar hour angles.
}
\description{
Compute the solar hour angle which is defined as the time difference from 
noon times 15 degrees per hour (i.e., negative during the morning and 
positive during the afternoon).
}
\examples{
## current solar hour angle
solarHourAngle(Sys.time(), unit = "degrees")

}

