% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarIncidenceAngle.R
\docType{methods}
\name{solarIncidenceAngle}
\alias{solarIncidenceAngle}
\alias{solarIncidenceAngle,POSIXt-method}
\alias{solarIncidenceAngle,Raster-method}
\alias{solarIncidenceAngle,character-method}
\title{Compute solar incidence angle}
\usage{
\S4method{solarIncidenceAngle}{POSIXt}(x, dem, formula = c("Cooper",
  "Spencer"), neighbors = 8L, unit = c("radians", "degrees"))

\S4method{solarIncidenceAngle}{character}(x, dem, formula = c("Cooper",
  "Spencer"), neighbors = 8L, unit = c("radians", "degrees"), ...)

\S4method{solarIncidenceAngle}{Raster}(x, origin = strptime("1993-01-01",
  "\%Y-\%m-\%d"), dem, formula = c("Cooper", "Spencer"), neighbors = 8L,
  unit = c("radians", "degrees"), ...)
}
\arguments{
\item{x}{A date-time object of class \code{POSIX*}, or a \code{character} 
object that can be coerced by \code{\link{strptime}}, or a \code{Raster*} 
object holding such information.}

\item{dem}{\code{RasterLayer}, digital elevation model in EPSG:4326
(see \url{http://spatialreference.org/ref/epsg/wgs-84/}).}

\item{formula}{\code{character}, formula from which to calculate the angular
declination of the sun, see \code{\link[satellite]{declination}}.}

\item{neighbors}{\code{integer}, number of focal cells from which to 
calculate surface slope and aspect, see \code{\link{terrain}}.}

\item{unit}{\code{character}, determines whether to return the solar 
incidence angle in radians (default) or degrees.}

\item{...}{If \code{x} is a \code{character} object, formatting arguments 
passed on to \code{\link{strptime}}, or if \code{x} is a \code{Raster*} 
object, formatting arguments related with the specified \code{origin}.}

\item{origin}{A \code{POSIX*} or \code{character} object that defines the 
origin of \code{x}, see \code{\link{solarHourAngle}}.}
}
\value{
A \code{Raster*} object with pixel-wise solar incidence angles.
}
\description{
Compute the solar incidence angle over non-planar surfaces following the
equation from Duffie and Beckman (2013).
}
\note{
An angle of more than 90 degrees, or likewise \code{pi/2}, indicates that the 
sun is behind the surface.
}
\examples{
## import sample data
dem <- system.file("extdata/DEM.TIF", package = "satellite")
dem <- trim(projectRaster(raster(dem), crs = "+init=epsg:4326"))

## current solar incidence angle
solarIncidenceAngle(Sys.time(), dem, formula = "Spencer", 
                    unit = "degrees")

}
\references{
Duffie JA, Beckman WA (2013) Solar Engineering of Thermal Processes. Wiley:
Hoboken, New Jersey, ISBN: 978-0-470-87366-3,
\url{http://eu.wiley.com/WileyCDA/WileyTitle/productCd-0470873663.html}.
}
\seealso{
\code{\link[satellite]{declination}}, \code{\link{solarHourAngle}}, 
\code{\link{terrain}}.
}

