% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetSat.R
\docType{methods}
\name{subsetSat}
\alias{subsetSat}
\title{Subset a Satellite object}
\usage{
\S4method{subsetSat}{Satellite}(x, subset)
}
\arguments{
\item{x}{vector of one or more satellite data files, raster::RasterStack}

\item{subset}{name of the computed instance (e.g. SC, cropped, REF_AtmosCorr etc.) or
row numbers of layers to be selected.}
}
\value{
Satellite object
}
\description{
Method to subset a Satellite Object. IMPORTANT: if name of target variable where subset is to be stored is the
same as the old one all other information will be lost!
}
\details{
Subsetting of computation stages.
}
\examples{
new_sat <- subsetSat(old_sat,subset = "cropped")
}

