% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfacePressure.R
\docType{methods}
\name{surfacePressure}
\alias{surfacePressure}
\alias{surfacePressure,Raster-method}
\alias{surfacePressure,numeric-method}
\title{Compute near-surface pressure}
\usage{
\S4method{surfacePressure}{Raster}(gp, ta, dem, p, ...)

\S4method{surfacePressure}{numeric}(gp, ta, dem, p)
}
\arguments{
\item{gp}{Geopotential heights as \code{Raster*} or, if a spatially distinct 
location is processed, as \code{numeric} vector.}

\item{ta}{Temperature profile of the same class as \code{gp}.}

\item{dem}{If \code{gp} is a \code{Raster*} object, a digital elevation model 
as \code{RasterLayer} or \code{numeric} vector (e.g. derived from 
\code{\link{getValues}}); else a single \code{numeric} elevation value.}

\item{p}{\code{numeric} pressure levels associated with \code{gp}.}

\item{...}{If \code{gp} is a \code{Raster*} object, further arguments passed 
on to \code{\link{writeRaster}}.}
}
\value{
If \code{gp} is a \code{Raster*} object, a \code{RasterLayer} with 
near-surface air pressure; else a single \code{numeric} value.
}
\description{
Compute the near-surface air pressure from the barometric formula using the
geopotential height of certain pressure levels, the referring temperature 
profile and the underlying digital elevation model as input data.
}
\examples{
### example from MYD07_L2 (20 levels) ---

## pressure levels             
p <- c(5, 10, 20, 30, 50, 70, 100, 150, 200, 250, 300, 400, 500, 620, 700, 
       780, 850, 920, 950, 1000) 
       
## geopotential heights
gp <- c(35005.573, 30329.573, 25786.936, 23204.563, 20074.706, 18113.499, 
        16109.901, 13765.185, 11966.857, 10484.984, 9218.005, 7115.921, 
        5399.528, 3679.846, 2676.816, 1760.488, 1018.150, NA, NA, NA)
        
## temperature profile         
ta <- c(234.6088, 226.9892, 220.2491, 214.4910, 203.8553, 194.2660, 191.1759, 
        205.9155, 220.9708, 232.4576, 241.8924, 256.9488, 267.5008, 277.0947, 
        283.7734, 289.8672, 294.1234, NA, NA, NA)   
    
## pixel elevation     
z <- 3625.708                                                    

## compute near-surface air pressure
surfacePressure(gp, ta, z, p)

}
\references{
Roedel W, Wagner T (2011) Physik unserer Umwelt: Die Atmosphaere. Springer: 
Heidelberg, Dordrecht, London, New York, ISBN: 978-3-642-15728-8. Available 
online at \url{http://cost733class.geo.uni-augsburg.de/moin/iguawiki/data/pages/KursmaterialWS1213_HsKlivar/attachments/Roedel_Die_Atmosphaere.pdf}.
}

