% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/surfaceReflectance.R
\name{surfaceReflectance}
\alias{surfaceReflectance}
\title{Compute surface reflectance using radiative transfer approximations}
\usage{
surfaceReflectance(bands_rad, coefs, ESun, Lp, model = "DOS2")
}
\arguments{
\item{coefs}{metadata from \code{\link{landsatMetadata}}}

\item{model}{to be used (DOS2, DOS4; must be the same as used by \
code{\ling{pathRadiance}})}

\item{sensor}{satellite sensor (Landsat 8/7/5/4)}

\item{date}{date of the satellite overpath}
}
\value{
Raster or RasterStack object containing surface reflectance data.
}
\description{
The function computes the surface reflectance using one of the following
approximation models:
- absolute radiance correction \cr
- DOS2: a dark object substraction model by Chavez (1996)
- DOS4: a dark object substratcion model by Moran et al. (1992)
}
\details{
The surface reflectance computation is based on a dark object
approach using either the DOS2 (Chavez 1996) or DOS4 (Moran et al. 1992)
model.

The minimum reflectance values for the dark object are identified using the
approximation of Chavez (1988, see \code{\link{pathRadiance}} for details).

The atmospheric transmittance towards the sensor (Tv) is approximated by
1.0 (DOS2, Chavez 1996) or rayleigh scattering (DOS4, Moran et al. 1992)

The atmospheric transmittance from the sun (Tz) is approximated by the
cosine of the sun zenith angle (DOS2, Chavez 1996) or again using rayleigh
scattering (DOS4, Moran et al. 1992).

The downwelling diffuse irradiance is approximated by 0.0 (DOS2, Chavez 1996)
or the hemispherical integral of the path radiance (DOS4, Moran et al. 1992).

Equations are taken from Song et al. (2001).
}
\examples{
not run:
surfaceReflectance(band_rad, bnbr, coefs, ESun, Lp)
}
\references{
Chavez Jr PS (1988) An improved dark-object subtraction technique
for atmospheric scattering correction of multispectral data. Remote Sensing
of Environment 24/3, doi:10.1016/0034-4257(88)90019-3, available online at
 \url{http://www.sciencedirect.com/science/article/pii/0034425788900193}

Chavez Jr PS (1996) Image-based atmospheric corrections revisited and
improved. Photogrammetric Engineering and Remote Sensing 62/9,
available online at
\url{http://www.asprs.org/PE-RS-Journals-1996/PE-RS-September-1996.html}

Goslee SC (2011) Analyzing Remote Sensing Data in R: The landsat
Package. Journal of Statistical Software,43/4, 1-25. URL
\url{http://www.jstatsoft.org/v43/i04/}.

Moran MS, Jackson RD, Slater PN, Teillet PM (1992) Evlauation of simplified
procedures for rretrieval of land surface reflectane factors from satellite
sensor output.Remote Sensing of Environment 41/2-3, 169-184,
doi:10.1016/0034-4257(92)90076-V,
URL \url{http://www.sciencedirect.com/science/article/pii/003442579290076V}.

Song C, Woodcock CE, Seto KC, Lenney MP, Macomber SA (2001) Classification
and Change Detection Using Landsat TM Data: When and How to Correct
Atmospheric Effects? Remote Sensing of Environment 75/2,
doi:10.1016/S0034-4257(00)00169-3, URL
\url{http://www.sciencedirect.com/science/article/pii/S0034425700001693}
}

