% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toaIrradianceTable.R
\name{toaIrradianceTable}
\alias{toaIrradianceTable}
\title{Get extraterrestrial solar irradiance (ESun) from readily tabulated values}
\usage{
toaIrradianceTable(sensor, normalize = TRUE, date)
}
\arguments{
\item{sensor}{sensor name ("Landsat 8/7")}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{date}{date of the sensor overpath (YYYY-MM-DD or POSIX* object) (
only required if ESun should be corrected to the actual earth sun distance)}
}
\value{
vector object containing ESun for each band
}
\description{
Get mean extraterrestrial solar irradiance (ESun) using publisch values.
}
\details{
Computation of ESun is taken from Updike and Comp (2011). Sun-earth
distance is computed using \code{\link{earthSun}}.
}
\examples{
toaIrradianceTable(sensor = "Landsat 7", normalize = FALSE, "2015-01-01")
}
\references{
Tabulated values of ESun are taken from

Landsat 7: Landsat7 handbook, tab 11.3 (Thuillier spectrum), URL
\url{http://landsathandbook.gsfc.nasa.gov/pdfs/Landsat7_Handbook.pdf}.

Landsat 4, 5: Chander G, Markham B (2003) Revised Landsat-5 TM radiometric
calibration procedures and postcalibration dynamic ranges.  IEEE Transaction
on Geoscience and Remote Sensing 41/11, doi:10.1109/LGRS.2007.898285, URL
\url{http://landsathandbook.gsfc.nasa.gov/pdfs/L5TMLUTIEEE2003.pdf}.
}
\seealso{
\code{\link{eSun}} for wrapping this function and alternative
derivation of ESun and \code{\link{earthSun}} for calculating the sun-earth
distance based on the day of the year which is called by this function if
ESun should be corrected for actual earth sun distance.
}

