% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alignGeometry.R
\docType{methods}
\name{alignGeometry}
\alias{alignGeometry}
\alias{alignGeometry,RasterLayer-method}
\alias{alignGeometry,RasterStack-method}
\title{Align raster geometry between two data sets}
\usage{
\S4method{alignGeometry}{Satellite}(x, template, band_codes, type,
  method = c("bilinear", "ngb"))

\S4method{alignGeometry}{RasterStack}(x, template, method = c("bilinear",
  "ngb"))

\S4method{alignGeometry}{RasterLayer}(x, template, method = c("bilinear",
  "ngb"))
}
\arguments{
\item{x}{Satellite or Raster* object to be resampled.}

\item{template}{Raster* or spatial data set from which geometry can be
extracted.}

\item{band_codes}{Band ID(s) to be resampled. If not supplied, all bands will
be considered for resampling.}

\item{type}{(Missing)}

\item{method}{Method for resampling; "bilinear" for bilinear interpolation
(default) or "ngb" for nearest neighbor interpolation. See e.g.
\code{\link{resample}}, \code{\link{projectRaster}}.}
}
\value{
Satellite object with path radiance for each band in the metadata
(W m-2 micrometer-1)

raster::RasterStack object with converted layers

raster::RasterLayer object with converted layer
}
\description{
Align raster data by bringing it in the same geometry and extent.
If the data set is not in the same projection as the template, the alignment
will be computed by reprojection. If the data has already the same
projection, the data set will be cropped and aggregated prior to resampling
in order to reduce computation time.




}
\examples{
## sample data
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)

## import as 'Satellite' object
sat <- satellite(files)

## align geometry
alignGeometry(sat, template = getSatDataLayer(sat, "B008n"),
               band_codes = "B001n")
}

