% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcEarthSunDist.R
\name{calcEartSunDist}
\alias{calcEartSunDist}
\title{Compute earth-sun distance based on day of the year.}
\usage{
calcEartSunDist(date, formula = "Spencer")
}
\arguments{
\item{date}{Date of the sensor overpath (YYYY-MM-DD or POSIX* object)}

\item{formula}{Formula providing author to be used (Spencer, Mather, ESA)}
}
\value{
Vector object containing earth-sun distancen in AU
}
\description{
Earth-sun distance is computed by one of several empirical formulas.
}
\details{
Computation of earth-sun distance using formulas provided by
Spencer (1971), Mather (2005) or ESA.
}
\examples{
calcEartSunDist(date = "2015-01-01", formula = "Spencer")
}
\references{
The formulas are taken from the following sources:

Spencer: Spencer JW (1971) Fourier series representation of the position of
the sun. Search 2/5. Taken from
\url{https://goo.gl/lhi9UI}.

Mather:  Paul M. Mather (2005) Computer Processing of Remotely-Sensed Images:
An Introduction. Wiley, ISBN: 978-0-470-02101-9,
\url{http://eu.wiley.com/WileyCDA/WileyTitle/productCd-0470021012.html}.

ESA: ESA Earth Observation Quality Control: Landsat frequently asked questions.

See also: Bird R, Riordan C (1984) Simple solar spectral model for direct and
diffuse irradiance on horizontal and tilted planes at the Earth's surface for
cloudless atmospheres. \url{http://www.nrel.gov/docs/legosti/old/2436.pdf}.
}

