% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcHistMatch.R
\name{calcHistMatch}
\alias{calcHistMatch}
\title{Illumination correction across scenes using histogram matching}
\usage{
calcHistMatch(x, target, mask, ttab = FALSE, minv = NULL, maxv = NULL,
  step = NULL)
}
\arguments{
\item{x}{A Satellite object or a raster::RasterStack or raster::RasterLayer
providing the source band(s) which should be adjusted.}

\item{target}{The target band as raster::RasterLayer}

\item{mask}{An invariant feature mask as raster::RasterLayer}

\item{ttab}{Logical, whether to return the transformation table}

\item{minv}{Lower bound of the possible range for transformation
(if not provided, the minimum of both layers)}

\item{maxv}{Upper bound of the possible range for transformation
(if not provided, the maximum of both layers)}

\item{by}{Step size used to build the new historgram
(if not provided, 1 for integer master layer, 0.01 for float master layer)}
}
\value{
if \code{ttab = FALSE} a RasterLayer, if \code{ttab = TRUE} a list
with components
\code{recode} the trasnsformation table used to match the histograms
\code{newraster} the transformed RasterLayer
}
\description{
This function adjusts the illumination of individual bands across two scenes
using a histogram match.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)


x = getSatDataLayer(sat, "B004n")
target = getSatDataLayer(sat, "B005n")
}
\references{
This function is taken and only slightly adapted from the histmatch
function of Sarah C. Goslee (2011). Analyzing Remote Sensing Data in R: The
landsat Package. Journal of Statistical Software,43(4), 1-25. URL
\url{http://www.jstatsoft.org/v43/i04/}.
}

