% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibLinear.R
\name{calibLinear}
\alias{calibLinear}
\title{Convert DNs to radiance, reflectance and/or brightness temperature.}
\usage{
calibLinear(band, mult, add, szen, k1, k2)
}
\arguments{
\item{band}{raster, rasterstack or data frame object of the sensor band}

\item{mult}{multiplicative coefficient for value transformation (i.e. slope)}

\item{add}{additive coefficient for value transformation (i.e. offset)}

\item{szen}{cosine of solar zenith angle}

\item{k1,k2}{temperature correction parameters}
}
\value{
Raster object with converted values
}
\description{
The function converts the digital numbers of a Lansat 8 level 1B/T file
to radiance (rad), relfectance (ref) and/or brightness temperature (bt) using
the calibration coefficients from the metadata file. The reflectance
conversion can additionaly be include a sun zenith correction.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

bcde <- "002n"

calibLinear(band = getSatDataLayer(sat, bcde),
            mult = getSatRADM(sat, bcde),
            add = getSatRADA(sat, bcde))
}
\references{
The conversion functions are taken from USGS' Landsat 8 manual
which is available online at
\url{http://landsat.usgs.gov/Landsat8_Using_Product.php}
}
\seealso{
\code{\link{radiometricCorrection}} for converions of the DNs
including scene-based atmospheric correction.
}

