% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cropSat.R
\docType{methods}
\name{cropSat}
\alias{cropSat}
\alias{cropSat,Satellite-method}
\title{Crop satellite object}
\usage{
\S4method{cropSat}{Satellite}(x, y, subset = TRUE)
}
\arguments{
\item{x}{Satellite object.}

\item{y}{\code{raster::Extent} object.}

\item{subset}{Logical; if \code{TRUE} (default), all layers but the cropped
ones are being dropped; if \code{FALSE}, cropped layers are appended to the
Satellite object.}
}
\value{
A Satellite object consisting of cropped layers only. If
\code{subset = FALSE}, a Satellite object with the cropped layers appended.
}
\description{
The function is a wrapper around the \code{raster::crop} function to
easily crop a Satellite object by an \code{raster::extent} object.
}
\details{
Crop layers of a Satellite object to the size of a given
\code{raster::extent} object.
}
\examples{
## sample data
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)

## geographic extent of georg-gassmann-stadium (utm 32-n)
ext_ggs <- raster::extent(484015, 484143, 5627835, 5628020)

## crop satellite object by specified extent
sat <- cropSat(sat, ext_ggs)
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for \code{raster::crop}.
}

