% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demTools.R
\docType{methods}
\name{demTools}
\alias{demTools}
\alias{demTools,RasterLayer-method}
\alias{demTools,Satellite-method}
\title{Compute terrain characteristics from digital elevation models}
\usage{
\S4method{demTools}{Satellite}(x, method = "hillShade", bcde = "DEM")

\S4method{demTools}{RasterLayer}(x, sunElev = NULL, sunAzim = NULL,
  method = "hillShade")
}
\arguments{
\item{x}{An object of type Satellite containing a DEM or a DEM provided as
RasterLayer.}

\item{method}{currently "slope", "aspect" and "hillshade" are implemented}

\item{bcde}{The name of the DEM layer in the satellite object.}

\item{sunElev}{if \code{method = "hillShade"}: The the elevation angle of the
sun in degrees.
See parameter angle in \code{\link{hillShade}}}

\item{sunAzim}{if \code{method = "hillShade"}: The sun azimuth angle in
degree. See parameter direction in \code{\link{hillShade}}}
}
\value{
Satellite object with added Layer of calculated terrain information

raster::RasterLayer object of calculated terrain information
}
\description{
Compute terrain characteristics from digital elevation models using the
raster::terrain function or raster::hillSahde function.




}
\examples{
\dontrun{
## Only works with a dem
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
sat <- addSatDataLayer(sat, method, data = DEM, info = NULL, in_bcde = NULL)
sat <- demTools(sat)
}
}

