% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{calibLinear}
\alias{calibLinearInverse}
\alias{deprecated}
\alias{satAtmosCorr}
\alias{satCalib}
\alias{satInvarFeatures}
\alias{satPathRadDOS}
\alias{satTOAIrrad}
\title{Deprecated functions}
\usage{
satCalib(x, convert = "all", szen_correction = "TRUE")

calibLinear(band, mult, add, szen, k1, k2)

satTOAIrrad(x, method = "Table", model = "MNewKur", normalize = TRUE, esd)

satPathRadDOS(x, atmos_model = "DOS2", esun_mode = "RadRef")

satAtmosCorr(x, atmos_model = "DOS2", esun_mode = "RadRef")

calibLinearInverse(band, ref_mult, ref_add, rad_mult, rad_add, szen)

satInvarFeatures(x)
}
\arguments{
\item{x}{Satellite object}

\item{convert}{Convert dataset to radiance/reflectance/temperature}

\item{szen_correction}{Compute sun zenith correction}

\item{band}{Band of the sensor}

\item{mult}{Multiplicative calibration coefficient}

\item{add}{Additive multiplication coefficient}

\item{szen}{Sun zenith angle}

\item{k1}{BTT calibration coefficient}

\item{k2}{BTT calibration coefficient}

\item{method}{Method used for computation}

\item{model}{Model used for computation}

\item{normalize}{Normalize earth-sun distance}

\item{esd}{Earth sun distance}

\item{atmos_model}{Atmospheric model to be used}

\item{esun_mode}{Earth sun distance computation method to be used}

\item{ref_mult}{Multiplicative reflection calibration coefficient}

\item{ref_add}{Additive reflection calibration coefficient}

\item{rad_mult}{Multiplicative radiance calibration coefficient}

\item{rad_add}{Additive radiance calibration coefficient}
}
\description{
The functions have been implemented in the very beginning of the package
development, mainly to be used within the scope of a remote sensing course at
Marburg University. To ensure that the scripts developed within this course
will still work after the next major revision, they are still part of this
package, but they will mainly just foreward the respective call to the
more up-to-date function.
}

