% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eSun.R
\name{eSun}
\alias{eSun}
\title{Get extraterrestrial solar irradiance (ESun) for satellite bands}
\usage{
eSun(sensor, tab = TRUE, normalize = TRUE, coefs, date)
}
\arguments{
\item{sensor}{sensor name ("Landsat 8/7/5/4")}

\item{tab}{use tabulated values (TRUE) or compute values based on rsr (FALSE)}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{coefs}{Landsat 8 metadata from \code{\link{landsatCoefficients}}
(Landsat 8 only, see details)}

\item{date}{date of the sensor overpath (YYYY-MM-DD or POSIX* object), only
relevant if normalize = FALSE}
}
\value{
vector object containing ESun for each band
}
\description{
Get extraterrestrial solar irradiance (ESun) from tabulated or computed
values. If values are computed, mean solar spectral data and the band
specifiv relative spectral response functions (rsr) are used.

For Landsat 8, no tabulated values are availabe. Instead (and if tab = TRUE)
ESun will be calculated based on the actual maximum radiance and reflection
given in the metadata file. Otherwise they will be computed using the
approach discribed above.
}
\details{
If eSun should be corrected for the actual earth sun distance, an
approximation of this distance is computed based on the day of by
\code{\link{earthSun}}. For Landsat 8, the respective earth sun distance is
taken from the metadata of the scene.

Instead of returning tabulated values for Landsat 8 which are not available
in the official handbook, \code{\link{toaIrradianceRadRef}} is used to
compute the actual eSun value based on the scene's metadata. By default,
the resulting actual ESun will be normalized to a mean earth sun distance
which is also taken from the scene's metadata.

If ESun should be computed (all sensors), \code{\link{toaIrradianceModel}}
will be called by this function.
}
\examples{
landsat8_metadatafile <-   system.file("extdata",
"LC81950252013188LGN00_MTL.txt", package = "satellite")
coefs8 <- landsatCoefficients(landsat8_metadatafile)
eSun(sensor = "Landsat 8", tab = TRUE, coefs = coefs8)
}
\references{
Tabulated values of ESun are taken from

Landsat 7: Landsat7 handbook, tab 11.3 (Thuillier spectrum), URL
\url{http://landsathandbook.gsfc.nasa.gov/pdfs/Landsat7_Handbook.pdf}.

Landsat 4, 5: Chander G, Markham B (2003) Revised Landsat-5 TM radiometric
calibration procedures and postcalibration dynamic ranges.  IEEE Transaction
on Geoscience and Remote Sensing 41/11, doi:10.1109/LGRS.2007.898285, URL
\url{http://landsathandbook.gsfc.nasa.gov/pdfs/L5TMLUTIEEE2003.pdf}.
}
\seealso{
This function can be used as a wrapper for
\code{\link{toaIrradianceModel}} for computing ESun based on tabulated and
\code{\link{toaIrradianceRadRef}} for computing ESun based on band data.
}

