% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSolarZenith.R
\docType{methods}
\name{fromSolarZenith}
\alias{fromSolarZenith}
\alias{fromSolarZenith,Raster-method}
\alias{fromSolarZenith,numeric-method}
\title{Compute hour angle and time from the solar zenith angle}
\usage{
\S4method{fromSolarZenith}{numeric}(theta, phi, n, formula = c("Cooper",
  "Spencer"))

\S4method{fromSolarZenith}{Raster}(theta, phi, n, formula = c("Cooper",
  "Spencer"), daytime = c("AM", "PM"), origin = strptime("1993-01-01",
  "\%Y-\%m-\%d"), ...)
}
\arguments{
\item{theta}{\code{numeric} or \code{Raster*} object, sun zenith angle in 
degrees (see Note).}

\item{phi}{\code{numeric} or \code{Raster*} object, latitude in EPSG:4326 
(see \url{http://spatialreference.org/ref/epsg/wgs-84/}).}

\item{n}{\code{Date} object from which to extract the day of the year (DOY), 
passed on to \code{\link[satellite]{declination}}.}

\item{formula}{\code{character}, formula from which to calculate the angular 
declination of the sun, see \code{\link[satellite]{declination}}.}

\item{daytime}{\code{character}, possible options are "AM" (i.e., morning 
scene) and "PM".}

\item{origin}{A \code{POSIX*} object that defines the origin of \code{x}, see 
\code{\link{solarHourAngle}}.}

\item{...}{If \code{theta} is a \code{Raster*} object, further arguments 
passed on to \code{\link{writeRaster}}.}
}
\description{
Compute the solar hour angle and the time difference from noon associated 
therewith from the sun zenith angle.
}
\note{
Negative and positive solar hour angles indicate solar zenith angles during 
the morning and afternoon, respectively. Note that the sun zenith angle can 
easily be derived from the elevation angle through 
\code{sun zenith = 1 - elevation}.
}
\examples{
theta <- 30 # 30 deg zenith angle
delta <- declination(Sys.Date(), "Spencer", "degrees") # angular declination
phi <- 0    # geographic latitude, i.e. equator

fromSolarZenith(theta, delta, phi) 

}
\references{
PVEducation (2016) Elevation angle. Available online at 
\url{http://www.pveducation.org/pvcdrom/properties-of-sunlight/elevation-angle}.
}
\seealso{
\code{\link[satellite]{declination}}, \code{\link{solarHourAngle}}.
}

