% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exwin.R
\name{incl}
\alias{incl}
\title{Compute similarity index}
\usage{
incl(x)
}
\arguments{
\item{pred,}{ref raster object containing the predicted and referen values respectivly.}

\item{win}{max window size. For now defaults to 3.}

\item{lweights}{vector with weights for layers of each window.}

\item{cweights}{matrix with weights for class changes.}
}
\value{
similarity index
}
\description{
Compute similarity index based on expanding window approach
}
\details{
Similarity index based on the approach in \code{\link{http://www.likbez.com/AV/PUBS/Kuhnert2005PERS.pdf}}.
}
\examples{
not run:
exwin(pred_raster, training_sites_shape)
}
\seealso{
\code{\link{compContingencyTable}} \code{\link{calcKappa}} for calculating kappa coefficients based on
contingency tables.
}

