% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite-package.R
\docType{data}
\name{l7}
\alias{l7}
\title{Landsat 7 sample data}
\format{\code{RasterStack} with bands 1-8 (incl. QA) of 41 by 41 pixels.}
\source{
\url{https://earthexplorer.usgs.gov/}
}
\description{
This dataset comes from the USGS. It contains part of the 
Landsat 7 scene LE07_L1TP_195025_20010730_20170204_01_T1 (Collection 1 
Level-1) from 2001-07-30 over Maburg, Germany.
}
\details{
Use of this data requires your agreement to the USGS regulations on 
using Landsat data.
}
\examples{
plotRGB(l7, r = 3, b = 1, stretch = "hist")
}
