% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite-package.R
\docType{data}
\name{l7}
\alias{l7}
\title{Landsat 7 sample data}
\format{\code{raster::RasterStack} with 8 bands of 41 by 41 pixels.}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\description{
This dataset comes from the USGS. It contains part of the
Landsat scene LE71950252001211EDC00 from 2001-07-30 over Maburg, Germany.
}
\details{


Use of this data requires your agreement to the USGS regulations on
using Landsat data.
}

