% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite-package.R
\docType{data}
\name{l8}
\alias{l8}
\title{Landsat 8 sample data}
\format{\code{raster::RasterStack} with 10 bands of 41 by 41 pixels.}
\source{
\url{http://earthexplorer.usgs.gov/}
}
\description{
This dataset comes from the USGS. It contains part of the
Landsat scene LC81950252013188LGN00 from 2013-07-07 over Maburg, Germany.
}
\details{


Use of this data requires your agreement to the USGS regulations on
using Landsat data.
}

