% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/landsatCoefficients.R
\name{landsatCoefficients}
\alias{landsatCoefficients}
\title{Get calibration information from Landsat 8 standard level 1B/T filename.}
\usage{
landsatCoefficients(filepath)
}
\arguments{
\item{filepath}{path and filename to the landsat metadata file}
}
\value{
Dataframe containing addtition and multiplication coefficients for
reflectance (REFA, REFM) and radiance (RADA, RADM), brightness temperature
correction parameters (BTK1, BTK2), sun elevation angle (SELEV),
sun zenith angle (SZEN) and sun azimuth angle (SAZM).
}
\description{
The function scans a Lansat 8 metadata file for calibration coefficients and
solar geometry information. Depending on the user selection, the calibration
coefficients are for reflectance or radiance.
}
\examples{
not run:
landsatCoefficients(filepath = "Name_of_Landsat_Metadata_File")
}

