% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metaFilePathLandsat.R
\name{metaFilePathLandsat}
\alias{metaFilePathLandsat}
\title{Get metadata filename from Landsat 8 standard level 1B/T filename.}
\usage{
metaFilePathLandsat(filepath)
}
\arguments{
\item{filepath}{path and filename to the landsat band file (not the metadata)}
}
\value{
vector containing band as character [1] and metadata filepath [2]
}
\description{
The function builds the metadata filename  based on standard level 1B/T
Landsat filenames.
}
\examples{
not run:
metaFilePathLandsat(filepath = "Name_of_Landsat_Band")
}

