% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite-package.R
\docType{package}
\name{satellite-package}
\alias{satellite-package}
\alias{satellitepackage}
\title{Smorgasboard for remote sensing functions.}
\description{
Smorgasboard for remote sensing functions
}
\details{
The package provides a variety of functions which are useful for 
handling, manipulating and visualizing remote sensing data.
}
\references{
Some functions are taken and/or adopted from Sarah C. Goslee 
(2011). Analyzing Remote Sensing Data in R: The landsat Package. Journal of 
Statistical Software, 43(4), 1-25. Online available at 
\url{http://www.jstatsoft.org/v43/i04/}.
}
\author{
Thomas Nauss, Hanna Meyer, Florian Detsch, Tim Appelhans \cr
\cr
\emph{Maintainer:} Florian Detsch \email{fdetsch@web.de}
}
\keyword{package}
