% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satellite.R
\docType{methods}
\name{satellite}
\alias{satellite}
\alias{satellite,RasterStack-method}
\alias{satellite,character-method}
\title{Create a Satellite object}
\usage{
\S4method{satellite}{character}(x, meta, log)

\S4method{satellite}{RasterStack}(x, meta, log)
}
\arguments{
\item{x}{A vector of one or more satellite data files or a
raster::RasterStack}

\item{meta}{Optional supply a metadata object (e.g. returned from
\code{\link{compMetaLandsat}}). If x is a satellite data file and recognised
as "Landsat", then the meta data is automatically extracted from the
respective meta information file if both the satellite data and the meta
data file follow the naming from the USGS Earth Explorer.
#'}

\item{log}{Optional supply a log entry.}
}
\value{
Satellite object
}
\description{
Method to create a Satellite Object




}
\details{
A satellite object consists of three data sections:
(i) a raster data section which holds the actual data values of the
respective sensor bands, (ii) a meta data grid which holds meta information
for each of the sensor band layers (e.g. calibration coefficients, type of
sensor band etc.) and (iii) a list of log information which records the
processing history of the entire data set.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
}
\seealso{
\code{\link{compMetaLandsat}} to get more information about the
structure of the metadata component.
}

