% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satInfo.R
\name{satInfo}
\alias{addSatDataLayer}
\alias{addSatLog}
\alias{addSatMetaParam}
\alias{getSatBCDE}
\alias{getSatBCDECalib}
\alias{getSatBCDESolar}
\alias{getSatBCDESolarCalib}
\alias{getSatBID}
\alias{getSatBTK1}
\alias{getSatBTK2}
\alias{getSatCalib}
\alias{getSatDATE}
\alias{getSatDataLayer}
\alias{getSatDataLayers}
\alias{getSatESD}
\alias{getSatESUN}
\alias{getSatLMAX}
\alias{getSatLMIN}
\alias{getSatLNBR}
\alias{getSatLog}
\alias{getSatMeta}
\alias{getSatMetaLayer}
\alias{getSatPRAD}
\alias{getSatParam}
\alias{getSatRADA}
\alias{getSatRADM}
\alias{getSatREFA}
\alias{getSatREFM}
\alias{getSatRadMax}
\alias{getSatRadMin}
\alias{getSatRefMax}
\alias{getSatRefMin}
\alias{getSatSAZM}
\alias{getSatSELV}
\alias{getSatSID}
\alias{getSatSZEN}
\alias{getSatSensor}
\alias{getSatSpectrum}
\alias{satInfo}
\title{Get or access Satellite object information used by various functions}
\usage{
getSatDataLayers(sat)

getSatDataLayer(sat, bcde)

addSatDataLayer(sat, bcde, data)

getSatMeta(sat)

getSatLog(sat)

addSatLog(sat, info = NA_character_, layers = NA_character_,
  output = NA_character_)

addSatMetaParam(sat, meta_param)

getSatParam(sat, param, bcde, return_bcde = TRUE)

getSatBCDE(sat)

getSatBID(sat)

getSatSID(sat)

getSatSensor(sat)

getSatSpectrum(sat)

getSatBCDESolar(sat)

getSatCalib(sat)

getSatBCDECalib(sat, bcde, id)

getSatBCDESolarCalib(sat, bcde, id)

getSatRadMax(sat, bcde)

getSatRadMin(sat, bcde)

getSatRefMax(sat, bcde)

getSatRefMin(sat, bcde)

getSatESD(sat)

getSatESUN(sat, bcde)

getSatSZEN(sat, bcde)

getSatSAZM(sat, bcde)

getSatSELV(sat, bcde)

getSatMetaLayer(sat, bcde)

getSatLNBR(sat, bcde)

getSatLMIN(sat, bcde)

getSatLMAX(sat, bcde)

getSatRADA(sat, bcde)

getSatRADM(sat, bcde)

getSatREFA(sat, bcde)

getSatREFM(sat, bcde)

getSatBTK1(sat, bcde)

getSatBTK2(sat, bcde)

getSatDATE(sat, bcde)

getSatPRAD(sat, bcde)
}
\arguments{
\item{sat}{Satellite object (see \code{\link{satellite}})}

\item{bcde}{band code}
}
\value{
Objects of respective type (see \code{\link{satellite}})
}
\description{
Get information from class Satellite.




}
\details{
The functions are generally self explaining in that sence that
\code{get*} returns the respective information and \code{set*} sets the
respective information from/in the Satellite object.

\code{addSatLog} adds a log entry to the Satellite object
}
\examples{
# List of input files
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

# Raster stack l8
sat <- satellite(l8)
}

