% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satInfo.R
\name{satInfo}
\alias{satESD}
\alias{satInfo}
\alias{satRadMax}
\alias{satRadMin}
\alias{satRefMax}
\alias{satRefMin}
\alias{satSID}
\alias{satSensor}
\title{Get or access Satellite object information used by various functions}
\usage{
satInfo(sat)

satSID(sat)

satSensor(sat)

satRadMax(sat)

satRadMin(sat)

satRefMax(sat)

satRefMin(sat)

satESD(sat)
}
\arguments{
\item{sat}{Satellite object (see \code{\link{satellite}})}
}
\value{
Objects of respective type (see \code{\link{satellite}})
}
\description{
Get information from class Satellite.
}
\details{
The functions above return the following information:
\itemize{
  \item \code{satESD} returns
  \item \code{satInfo} returns the entire metadata
  \item \code{satRadMax} returns the maximum radiance in band(s)
  \item \code{satRadMin} returns the minimum radiance in band(s)
  \item \code{satRefMax} returns the maximum reflectance in band(s)
  \item \code{satRefMin} returns the minimum reflectance in band(s)
  \item \code{satSensor} returns the sensor name
  \item \code{satSID} returns the sensor ID
}
}
\section{Functions}{
\itemize{
\item \code{satSID}: 

\item \code{satSensor}: 

\item \code{satRadMax}: 

\item \code{satRadMin}: 

\item \code{satRefMax}: 

\item \code{satRefMin}: 

\item \code{satESD}: 
}}
\examples{
satInfo()
}

