% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satInfo.R
\name{satInfo}
\alias{addRasterMeta2Sat}
\alias{addSatDataLayer}
\alias{addSatLog}
\alias{addSatMetaEntry}
\alias{addSatMetaParam}
\alias{countSatDataLayers}
\alias{createRasterMetaData}
\alias{createSatBCDE}
\alias{getSatBCDE}
\alias{getSatBCDECalib}
\alias{getSatBCDEFromSpectrum}
\alias{getSatBCDEFromType}
\alias{getSatBCDESolar}
\alias{getSatBCDESolarCalib}
\alias{getSatBCDESres}
\alias{getSatBCDEThermal}
\alias{getSatBCDEThermalCalib}
\alias{getSatBCDEType}
\alias{getSatBID}
\alias{getSatBTK1}
\alias{getSatBTK2}
\alias{getSatBandInfo}
\alias{getSatCalib}
\alias{getSatDATE}
\alias{getSatDataLayer}
\alias{getSatDataLayers}
\alias{getSatESD}
\alias{getSatESUN}
\alias{getSatLMAX}
\alias{getSatLMIN}
\alias{getSatLNBR}
\alias{getSatLayerfromData}
\alias{getSatLog}
\alias{getSatMeta}
\alias{getSatMetaBCDETemplate}
\alias{getSatMetaLayer}
\alias{getSatPRAD}
\alias{getSatParam}
\alias{getSatProjection}
\alias{getSatRADA}
\alias{getSatRADM}
\alias{getSatREFA}
\alias{getSatREFM}
\alias{getSatRadMax}
\alias{getSatRadMin}
\alias{getSatRefMax}
\alias{getSatRefMin}
\alias{getSatRes}
\alias{getSatSAZM}
\alias{getSatSELV}
\alias{getSatSID}
\alias{getSatSZEN}
\alias{getSatSensor}
\alias{getSatSensorGroup}
\alias{getSatSensorInfo}
\alias{getSatSpectrum}
\alias{getSatType}
\alias{getSatXRes}
\alias{getSatYRes}
\alias{satInfo}
\alias{setSatBCDE}
\title{Get or access Satellite object information used by various functions}
\usage{
getSatDataLayers(sat, bcde = NULL)

getSatDataLayer(sat, bcde)

getSatMeta(sat, bcde)

getSatMetaBCDETemplate(sat, bcde)

getSatLog(sat)

setSatBCDE(sat, bcde)

createSatBCDE(sat, width = 3, flag = 0, prefix = "B", postfix = "n")

addSatMetaParam(sat, meta_param)

addSatMetaEntry(sat, meta_param)

addSatLog(sat, info = NA_character_, in_bcde = NA_character_,
  out_bcde = NA_character_)

addSatDataLayer(sat, bcde, data, meta_param, info, in_bcde)

addRasterMeta2Sat(sat)

createRasterMetaData(rst)

countSatDataLayers(sat)

getSatParam(sat, param, bcde, return_bcde = TRUE)

getSatBCDE(sat, lnbr)

getSatBID(sat, bcde)

getSatSID(sat)

getSatSensor(sat)

getSatSensorGroup(sat)

getSatSensorInfo(sat)

getSatSpectrum(sat, bcde)

getSatBCDESolar(sat)

getSatBCDEThermal(sat)

getSatXRes(sat, bcde)

getSatYRes(sat, bcde)

getSatRes(sat, bcde)

getSatType(sat, bcde)

getSatCalib(sat, bcde)

getSatBCDEType(sat, bcde, id)

getSatBCDEFromType(sat, id = "VIS")

getSatBCDEFromSpectrum(sat, spectrum = "solar")

getSatBCDESres(sat, bcde, id)

getSatBCDECalib(sat, bcde, id)

getSatBCDESolarCalib(sat, bcde, id)

getSatBCDEThermalCalib(sat, bcde, id)

getSatBandInfo(sat, bcde, return_calib = TRUE)

getSatRadMax(sat, bcde)

getSatRadMin(sat, bcde)

getSatRefMax(sat, bcde)

getSatRefMin(sat, bcde)

getSatESD(sat)

getSatESUN(sat, bcde)

getSatSZEN(sat, bcde)

getSatSAZM(sat, bcde)

getSatSELV(sat, bcde)

getSatMetaLayer(sat, bcde)

getSatLayerfromData(sat, bcde, nbr)

getSatLNBR(sat, bcde)

getSatLMIN(sat, bcde)

getSatLMAX(sat, bcde)

getSatRADA(sat, bcde)

getSatRADM(sat, bcde)

getSatREFA(sat, bcde)

getSatREFM(sat, bcde)

getSatBTK1(sat, bcde)

getSatBTK2(sat, bcde)

getSatDATE(sat, bcde)

getSatPRAD(sat, bcde)

getSatProjection(sat, bcde)
}
\arguments{
\item{sat}{Satellite object (see \code{\link{satellite}}).}

\item{bcde}{Band code.}

\item{meta_param}{Metadata parameters used to document new data layer}

\item{info}{Log information added to metadata}

\item{in_bcde}{BCDE of layer used as input dataset}

\item{out_bcde}{BCDE of layer used as output dataset}

\item{data}{Data layer of a satellite object}

\item{rst}{Input raster::Raster* object from which to extract metadata.}

\item{param}{Parameter of the metadata set (i.e. colname)}

\item{return_bcde}{Return bcde as attribute (TRUE/FALSE)}

\item{lnbr}{Layer number}

\item{id}{Type of the sensor band}

\item{spectrum}{Spectral region, e.g. "solar" or "thermal".}

\item{return_calib}{Return calibration information (TRUE/FALSE)}

\item{nbr}{Return specific data layer selected by number}
}
\value{
Objects of respective type (see \code{\link{satellite}}).
}
\description{
Get information from class Satellite.




}
\details{
The functions are generally self-explaining in that sence that
\code{get*} returns the respective information and \code{set*} sets the
respective information from/in the Satellite object.

\code{addSatLog} adds a log entry to the Satellite object.
}
\examples{
# List of input files
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

# Raster stack l8
sat <- satellite(l8)
}

