% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/satInfo.R
\name{satInfo}
\alias{addSatLog}
\alias{getSatESD}
\alias{getSatLayers}
\alias{getSatLog}
\alias{getSatMeta}
\alias{getSatRadMax}
\alias{getSatRadMin}
\alias{getSatRefMax}
\alias{getSatRefMin}
\alias{getSatSID}
\alias{getSatSensor}
\alias{satInfo}
\title{Get or access Satellite object information used by various functions}
\usage{
getSatLayers(sat)

getSatMeta(sat)

getSatLog(sat)

addSatLog(sat, info = NA_character_, layers = NA_character_,
  output = NA_character_)

getSatSID(sat)

getSatSensor(sat)

getSatRadMax(sat)

getSatRadMin(sat)

getSatRefMax(sat)

getSatRefMin(sat)

getSatESD(sat)
}
\arguments{
\item{sat}{Satellite object (see \code{\link{satellite}})}
}
\value{
Objects of respective type (see \code{\link{satellite}})
}
\description{
Get information from class Satellite.
}
\details{
The functions are generally self explaining in that sence that
\code{get*} returns the respective information and \code{set*} sets the
respective information from/in the Satellite object.

\code{addSatLog} adds a log entry to the Satellite object
}
\examples{
# List of input files
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)

# Raster stack l8
sat <- satellite(l8)
}

