% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satTOAIrrad.R
\docType{methods}
\name{satTOAIrrad}
\alias{satTOAIrrad}
\title{Get TOA solar irradiance (ESun) for the layers of a Satellite object}
\usage{
\S4method{satTOAIrrad}{Satellite}(x, method = "Table", model = "MNewKur",
  normalize = TRUE, esd)
}
\arguments{
\item{x}{object of type Satellite}

\item{method}{name of the method to be used ("Table", "Model", "RadRef)}

\item{model}{name of the model to be used if method is "Model"
(see \code{\link{calcTOAIrradModel}})}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{esd}{earth-sun distance (AU), if not supplied and necessary for
normalize, it is tried to take it from the metadata, otherwise it is estimated
by the day of the year using \code{\link{calcEartSunDist}}.}
}
\value{
Satellite object with ESun for each band in the metadata
}
\description{
Get extraterrestrial solar irradiance (ESun) from tabulated or computed
values. If values are computed, either (i) the mean solar spectral data and
the band specifiv relative spectral response functions (rsr) are used or (ii)
actual maximum radiance and reflection values are taken.
}
\details{
Tabulated values of ESun are compiled using \code{\link{calcTOAIrradRadTable}}.

Actual values of ESun are computed using \code{\link{calcTOAIrradRadRef}}.

Modeled values of ESun are computed using \code{\link{calcTOAIrradModel}}.

If eSun should be corrected for the actual earth sun distance, the distance
is taken from (i) the function parameter, (ii) the metadata or (iii)
approximation using \code{\link{calcEartSunDist}} which is called from
within the functions above.

Please refer to the respective functions for details on the computation.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LE7*.tif"), full.names = TRUE)
sat <- satellite(files)
satTOAIrrad(sat, method = "Table")

path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
satTOAIrrad(sat, method = "Model")

satTOAIrradModel(sat)
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
sat <- satellite(files)
satTOAIrrad(sat, method = "RadRef")
}
\references{
Please refer to the respective functions for references.
}
\seealso{
This function is a wrapper for
\code{\link{calcTOAIrradRadTable}}, \code{\link{calcTOAIrradRadRef}} and
\code{\link{calcTOAIrradModel}}.
}

