% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarIncidenceAngle.R
\name{solarIncidenceAngle}
\alias{hourAngle}
\alias{solarIncidenceAngle}
\title{Compute solar incidence and hour angle}
\usage{
solarIncidenceAngle(x, dem, formula = c("Cooper", "Spencer"),
  neighbors = 8L, unit = c("radians", "degrees"), ...)

hourAngle(x, ...)
}
\arguments{
\item{x}{A date-time object of class \code{POSIX*} or, alternatively, a
\code{character} object that can be coerced by \code{\link{strptime}}.}

\item{dem}{\code{RasterLayer}, digital elevation model in EPSG:4326
\url{http://spatialreference.org/ref/epsg/wgs-84/}.}

\item{formula}{Formula to deploy for the calculation of the angular
declination of the sun, see \code{\link[satellite]{declination}}.}

\item{neighbors}{\code{integer}, number of focal cells to use for the
computation of surface slope and aspect, see \code{\link{terrain}}.}

\item{unit}{\code{character}, determines whether the solar incidence angle is 
returned in radians (default) or degrees.}

\item{...}{If \code{x} is a \code{character} object, further arguments passed
on to \code{\link{strptime}}.}
}
\value{
\itemize{
\item \emph{solarIncidenceAngle}: A \code{RasterLayer} with pixel-wise solar 
incidence angles.
\item \emph{hourAngle}: A \code{numeric} solar hour angle.
}
}
\description{
Compute the solar incidence angle over non-planar surfaces following the
equation from Duffie and Beckman (2013) and the solar hour angle defined as 
the time difference from noon times 15 degrees per hour.
}
\examples{
## import sample data
dem <- system.file("extdata/DEM.TIF", package = "satellite")
dem <- trim(projectRaster(raster(dem), crs = "+init=epsg:4326"))

## current solar incidence angle
solarIncidenceAngle(Sys.time(), dem, formula = "Spencer", 
                    unit = "degrees")

## current solar hour angle
hourAngle(Sys.time())

}
\references{
Duffie JA, Beckman WA (2013) Solar Engineering of Thermal Processes. Wiley:
Hoboken, New Jersey, ISBN: 978-0-470-87366-3,
\url{http://eu.wiley.com/WileyCDA/WileyTitle/productCd-0470873663.html}.
}
\seealso{
\code{\link[satellite]{declination}}.
}

