% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subset}
\alias{[[,Satellite,ANY,ANY-method}
\alias{subset}
\alias{subset,Satellite-method}
\title{Subset Satellite object data layers}
\usage{
\S4method{subset}{Satellite}(x, sid, cid)

\S4method{[[}{Satellite,ANY,ANY}(x, i)
}
\arguments{
\item{x}{Satellite or raster::Raster* object providing the source band(s) to
be adjusted.}

\item{sid}{Band numbers or bcde which should be extracted}

\item{cid}{Calibration information used for subsetting (only works if sid is
not supplied to the function)}

\item{i}{Layer index(es) for subsetting.}
}
\value{
A Satellite object

A Satellite object

A Satellite object
}
\description{
This function subsets a Satellite object and returns the extracted dataset
as Satellite object.




}
\examples{
## sample data
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.TIF"), full.names = TRUE)
sat <- satellite(files)

sat[[2:5]]
subset(sat, cid = "SC")
}

