% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toaIrradianceModel.R
\name{toaIrradianceModel}
\alias{toaIrradianceModel}
\title{Compute extraterrestrial solar irradiance (ESun) for satellite bands}
\usage{
toaIrradianceModel(sensor, model = "MNewKur", normalize = TRUE, date)
}
\arguments{
\item{sensor}{sensor name ("Landsat 8/7")}

\item{model}{tabulated solar radiation model to be used}

\item{normalize}{normalize ESun to mean earth sun distance}

\item{date}{date of the sensor overpath (YYYY-MM-DD or POSIX* object) (
only required if ESun should be corrected to the actual earth sun distance)}
}
\value{
vector object containing ESun for each band
}
\description{
Compute mean extraterrestrial solar irradiance (ESun) using tabulated mean
solar spectral data and the band specifiv relative spectral response
functions (rsr).
}
\details{
Computation of ESun is taken from Updike and Comp (2011). Sun-earth
distance is computed using \code{\link{earthSun}}.
}
\examples{
toaIrradianceModel(sensor = "Landsat 7", model = "MNewKur")
}
\references{
Computation of ESun is taken from Updike T, Comp C (2011)
Radiometric use of WorldView-2 imagery. Technical Note, URL
\URL{https://www.digitalglobe.com/sites/default/files/Radiometric_Use_of_WorldView-2_Imagery(1).pdf}.

Tabulated relative spectral response functions (nm-1) are taken from the USGS
at \url{http://landsat.usgs.gov/instructions.php}.

Tabulated solar irradiance (W m-2 nm-1) are taken from the National Renewable
Energy Laboratory at \url{http://rredc.nrel.gov/solar/spectra/am0/modtran.html}.
}
\seealso{
\code{\link{eSun}} for wrapping this function and alternative
derivation of ESun and \code{\link{earthSun}} for calculating the sun-earth
distance based on the day of the year which is called by this function if
ESun should be corrected for actual earth sun distance.
}

