% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TopoCorr.R
\docType{methods}
\name{TopoCorr}
\alias{TopoCorr}
\alias{TopoCorr,RasterLayer-method}
\alias{TopoCorr,RasterStack-method}
\title{Correct for topographic effects}
\usage{
\S4method{TopoCorr}{RasterStack}(x, hillsh, cloudmask = NULL)

\S4method{TopoCorr}{RasterLayer}(x, hillsh, cloudmask = NULL)
}
\arguments{
\item{x}{Object of class satellite}

\item{cloudmask}{Logical. If TRUE then the cloudmask from the
satellite object (if available) will be considered in the regression model.}

\item{dem}{Object of class RasterLayer.
If not specified, an srtm elevation model will be downloaded.}
}
\value{
Satellite object with added, topographic corrected layers

Satellite object with added converted layers

raster::RasterStack object with converted layers

raster::RasterLayer object with converted layer
}
\description{
Correct for topographic effects




}
\details{
The method of Civco (1989) is applied:
First, an analytical hillshade image is created based on a DEM and sun elevation and
sun zenith information from the metadata. A regression between
the hillshade (independent variable) and each channel is then calculated
with consideration of a cloudmask (if available).
The regression coefficents are used to calibrate the hillshade raster
(for each channel individually).
Finally the calibrated hillshade image is substracted from the corresponding
channel and the mean value of the channel is added.
}
\examples{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
x<-satellite(files)
TopoCorr(x)
}
\references{
CIVCO, D.L. (1989): Topographic normalization of Landsat Thematic
Mapper digitalimagery.In: Photogrammetric Engineering & Remote Sensing, 55,
S. 1303–1309.
}

