% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TopoCorr.R
\docType{methods}
\name{TopoCorr}
\alias{TopoCorr}
\alias{TopoCorr,RasterLayer-method}
\alias{TopoCorr,RasterStack-method}
\alias{TopoCorr,Satellite-method}
\title{Correct for topographic effects.}
\usage{
\S4method{TopoCorr}{Satellite}(x, mask = TRUE)

\S4method{TopoCorr}{RasterStack}(x, hillsh, cloudmask = NULL)

\S4method{TopoCorr}{RasterLayer}(x, hillsh, cloudmask = NULL)
}
\arguments{
\item{x}{Satellite object.}

\item{mask}{Logical. If \code{TRUE}, the cloudmask from the Satellite object
(if available) will be considered in the regression model.}

\item{hillsh}{A \code{raster::RasterLayer} created with
\code{\link{raster::hillShade}}.}

\item{cloudmask}{A \code{raster::RasterLayer} in which clouds are masked with
NA values.}
}
\value{
If x is a Satellite object, a Satellite object with added,
topographic corrected layers; if x is a \code{raster::Raster*} object, a
\code{raster::Raster*} object with converted layer(s).
}
\description{
Correct for topographic effects.






}
\details{
The method of Civco (1989) is applied on atmospherically corrected bands
(if not already available in the Satellite object,
\code{\link{calcAtmosCorr}} is performed with its default settings.):
First, an analytical hillshade image is created based on a DEM and sun
elevation and sun zenith information from the metadata. A regression between
the hillshade (independent variable) and each channel is then calculated
with consideration of a cloudmask (if available).
The regression coefficents are used to calibrate the hillshade raster
(for each channel individually).
Finally, the calibrated hillshade image is subtracted from the corresponding
channel and the mean value of the channel is added.
}
\examples{
#' \dontrun{
path <- system.file("extdata", package = "satellite")
files <- list.files(path, pattern = glob2rx("LC8*.tif"), full.names = TRUE)
x <- satellite(files)
TopoCorr(x)
}
}
\references{
CIVCO, D.L. (1989): Topographic normalization of Landsat Thematic
Mapper digitalimagery. \emph{Photogrammetric Engineering & Remote Sensing},
55, 1303-1309.
}

